Bind new-design paint to a host FOCUS artboard and tighten editor chrome.
Require create_frame before content on Smart/auto, keep host FOCUS as the create target, hide the selection toolbar while side tools are open, and align video drag underlay with image layers.
