--- apps/api/app/services/design/prompts/skill_store.py
@@ -1,3011 +1,0 @@
-"""Runtime skills for lc_design — pluggable catalog + need_skills + triggers + files.
-
-Namespaces (conflict isolation):
-  - core  — system seed skills (source=seed); bare keys kept for BC
-  - ext   — file packs under ``.agents/skills`` + data/design_skills (source=file)
-  - user  — admin / user-extension skills (source=admin); keys use ``user.<local>``
-
-Also:
-  - ACL: preferred_tools + allowed_resources; user skills never reopen unrestricted surface
-  - I/O: pack/admin meta validation; need_skills pin/args vs input_schema; ops vs output_schema
-  - Version: integer bump + pack_version + design_skill_revision snapshots; pin via ``key@N``
-  - Hot reload: optional mtime watcher on seed JSON + file packs
-  - Formats: product ``_meta.json``+``SKILL.md``, or ``SKILL.md`` with YAML frontmatter
-"""
-from __future__ import annotations
-
-import io
-import json
-import logging
-import re
-import shutil
-import tempfile
-import threading
-import time
-import zipfile
-from pathlib import Path
-from typing import Any
-
-from sqlmodel import Session
-
-from app import crud
-from app.core import db as core_db
-
-logger = logging.getLogger(__name__)
-
-_SKILLS_READY = False
-_SKILLS_LOCK = threading.RLock()
-_HOT_RELOAD_STOP = threading.Event()
-_HOT_RELOAD_THREAD: threading.Thread | None = None
-_DISK_SIGNATURE: str | None = None
-
-SOURCE_SEED = "seed"
-SOURCE_ADMIN = "admin"
-SOURCE_FILE = "file"
-_PROTECTED_FROM_SEED = frozenset({SOURCE_ADMIN, SOURCE_FILE})
-_PROTECTED_FROM_FILE = frozenset({SOURCE_ADMIN, SOURCE_SEED})
-
-NS_CORE = "core"
-NS_EXT = "ext"
-NS_USER = "user"
-_VALID_NAMESPACES = frozenset({NS_CORE, NS_EXT, NS_USER})
-_SOURCE_TO_NS = {
-    SOURCE_SEED: NS_CORE,
-    SOURCE_FILE: NS_EXT,
-    SOURCE_ADMIN: NS_USER,
-}
-
-# Retired need_* prompt packs (methodology/vision/aesthetics → Skills). DELETE leftovers on sync.
-RETIRED_NEED_PROMPT_KINDS = frozenset({"design_spec", "vision", "aesthetics"})
-
-# Ops always allowed even when preferred_tools allowlist is active.
-_ALWAYS_ALLOW_OPS = frozenset(
-    {
-        "ask_user",
-        "update_node",
-        "delete_nodes",
-        "align_nodes",
-        "distribute_nodes",
-        "move_nodes",
-        "resize_nodes",
-        "reorder_nodes",
-        "group_nodes",
-        "ungroup_nodes",
-    }
-)
-
-MAX_SKILL_DETAIL_CHARS = 14000
-_META_NAMES = ("_meta.json", "meta.json")
-_SKILL_MD_NAMES = ("SKILL.md", "skill.md")
-_NS_KEY_RE = re.compile(r"^(core|ext|user)[.:/](.+)$", re.IGNORECASE)
-_PIN_RE = re.compile(r"^(.+?)@([0-9]+(?:\.[0-9]+){0,2})$")
-
-
-def _skill_md_path(pack_dir: Path) -> Path | None:
-    for name in _SKILL_MD_NAMES:
-        p = pack_dir / name
-        if p.is_file():
-            return p
-    return None
-
-
-def _unquote_yaml_scalar(raw: str) -> str:
-    s = str(raw or "").strip()
-    if len(s) >= 2 and s[0] == s[-1] and s[0] in ("'", '"'):
-        return s[1:-1].strip()
-    return s
-
-
-def _split_skill_md_frontmatter(text: str) -> tuple[dict[str, Any], str]:
-    """Parse ``SKILL.md`` YAML frontmatter (flat keys only).
-
-    Returns ``(meta, body)``. No frontmatter → ``({}, original_text)``.
-    """
-    raw = str(text or "").lstrip("\ufeff")
-    if not raw.startswith("---"):
-        return {}, raw.strip()
-    lines = raw.splitlines()
-    if not lines or lines[0].strip() != "---":
-        return {}, raw.strip()
-    end = -1
-    for i in range(1, len(lines)):
-        if lines[i].strip() == "---":
-            end = i
-            break
-    if end < 1:
-        return {}, raw.strip()
-
-    meta: dict[str, Any] = {}
-    key: str | None = None
-    buf: list[str] = []
-    fold = False
-
-    def flush() -> None:
-        nonlocal key, buf, fold
-        if not key:
-            buf = []
-            fold = False
-            return
-        if fold:
-            meta[key] = " ".join(x.strip() for x in buf if x.strip()).strip()
-        elif buf:
-            meta[key] = _unquote_yaml_scalar("\n".join(buf).strip())
-        else:
-            meta[key] = ""
-        key = None
-        buf = []
-        fold = False
-
-    for line in lines[1:end]:
-        if key and (line.startswith("  ") or line.startswith("\t")):
-            buf.append(line.strip())
-            continue
-        m = re.match(r"^([A-Za-z0-9_-]+)\s*:\s*(.*)$", line)
-        if not m:
-            if key:
-                buf.append(line.strip())
-            continue
-        flush()
-        key = m.group(1).strip()
-        rest = m.group(2).rstrip()
-        if rest in (">", "|", ">-", "|-"):
-            fold = True
-            buf = []
-        elif rest == "":
-            fold = True
-            buf = []
-        else:
-            fold = False
-            buf = [rest]
-    flush()
-
-    body = "\n".join(lines[end + 1 :]).strip()
-    return meta, body
-
-
-def _meta_from_agent_skill_frontmatter(
-    fm: dict[str, Any], *, folder: str
-) -> dict[str, Any] | None:
-    """Build a minimal pack meta from ``SKILL.md`` YAML frontmatter."""
-    if not isinstance(fm, dict) or not fm:
-        return None
-    name = str(fm.get("name") or folder or "").strip()
-    description = str(fm.get("description") or "").strip()
-    if not name and not description:
-        return None
-    key = _slug_local_key(name or folder) or folder
-    when = str(fm.get("when_to_use") or fm.get("whenToUse") or description).strip()
-    display = str(
-        fm.get("displayName") or fm.get("display_name") or fm.get("title") or name or key
-    ).strip() or key
-    return {
-        "skill_key": str(fm.get("skill_key") or fm.get("skillKey") or key).strip() or key,
-        "name": key,
-        "displayName": display,
-        "description": description or when,
-        "when_to_use": when,
-        "category": str(fm.get("category") or "agent").strip() or "agent",
-        "version": fm.get("version") or "1.0.0",
-        "scenes": "all",
-        "preferred_tools": [],
-        "triggers": [],
-        "locales": {
-            "en": {"displayName": display, "description": description or when},
-        },
-    }
-
-_RUNTIME_SKILL_KEYS: frozenset[str] | None = None
-_RUNTIME_SKILL_INDEX: dict[str, dict[str, Any]] | None = None
-
-
-def _skills_seed_path() -> Path:
-    from app.core.config import resolve_data_file
-
-    return resolve_data_file("design_skills_seed.json")
-
-
-def _repo_root() -> Path:
-    """``apps/api`` → repository root."""
-    from app.core.config import _API_ROOT
-
-    return _API_ROOT.parent.parent
-
-
-def _agents_skills_dir() -> Path:
-    """Repo-root official ext packs: ``<repo>/.agents/skills``."""
-    return _repo_root() / ".agents" / "skills"
-
-
-def _file_skills_dir() -> Path:
-    """Primary design_skills dir under apps/api/data/."""
-    from app.core.config import resolve_data_dir
-
-    return resolve_data_dir("design_skills")
-
-
-def _file_skills_dirs() -> list[Path]:
-    """``.agents/skills`` + ``apps/api/data/design_skills``."""
-    from app.core.config import api_data_dir
-
-    dirs: list[Path] = []
-    seen: set[str] = set()
-    for root in (
-        _agents_skills_dir(),
-        api_data_dir() / "design_skills",
-    ):
-        try:
-            resolved = root.resolve()
-        except Exception:
-            continue
-        key = str(resolved).lower()
-        if key in seen or not root.is_dir():
-            continue
-        seen.add(key)
-        dirs.append(root)
-    return dirs
-
-
-def _pack_has_product_meta(pack_dir: Path) -> bool:
-    return any((pack_dir / name).is_file() for name in _META_NAMES)
-
-
-def _load_skills_seed() -> list[dict[str, Any]]:
-    path = _skills_seed_path()
-    try:
-        parsed = json.loads(path.read_text(encoding="utf-8"))
-    except Exception:
-        return []
-    if not isinstance(parsed, dict):
-        return []
-    items = parsed.get("items") or []
-    if not isinstance(items, list):
-        return []
-    return [x for x in items if isinstance(x, dict) and str(x.get("skill_key") or "").strip()]
-
-
-_SEED = _load_skills_seed()
-_SEED_BY_KEY: dict[str, dict[str, Any]] = {
-    str(item.get("skill_key") or "").strip(): item for item in _SEED
-}
-_CORE_RESERVED_KEYS = frozenset(_SEED_BY_KEY.keys())
-
-
-def _normalize_namespace(raw: Any, *, source: str | None = None) -> str:
-    s = str(raw or "").strip().lower()
-    if s in _VALID_NAMESPACES:
-        return s
-    if source:
-        return _SOURCE_TO_NS.get(_normalize_source(source), NS_USER)
-    return NS_USER
-
-
-def split_namespace_key(raw: str) -> tuple[str | None, str]:
-    """Return (namespace|None, local_key) from ``core.x`` / ``ext:x`` / bare ``x``."""
-    s = str(raw or "").strip().lower()
-    if not s:
-        return None, ""
-    m = _NS_KEY_RE.match(s)
-    if m:
-        return m.group(1).lower(), str(m.group(2) or "").strip()
-    return None, s
-
-
-def qualify_skill_key(namespace: str, local_key: str) -> str:
-    ns = _normalize_namespace(namespace)
-    _, local = split_namespace_key(local_key)
-    local = local.strip().lower()
-    if not local:
-        return ""
-    if ns == NS_CORE:
-        # Core keeps bare storage keys for backward compatibility.
-        return local
-    return f"{ns}.{local}"
-
-
-def skill_kind_for_namespace(namespace: str) -> str:
-    return "core" if _normalize_namespace(namespace) == NS_CORE else "extension"
-
-
-def parse_skill_pin(raw: str) -> tuple[str, int | None, str | None]:
-    """Split ``key@2`` / ``key@1.0.0`` → (key, int_version|None, pack_version|None)."""
-    s = str(raw or "").strip()
-    if not s:
-        return "", None, None
-    m = _PIN_RE.match(s)
-    if not m:
-        return s, None, None
-    base = str(m.group(1) or "").strip()
-    pin = str(m.group(2) or "").strip()
-    if re.fullmatch(r"\d+", pin):
-        return base, max(1, int(pin)), None
-    return base, None, pin
-
-
-def _parse_json_object(raw: Any) -> dict[str, Any] | None:
-    if raw is None or raw is False:
-        return None
-    if isinstance(raw, dict):
-        return raw
-    s = str(raw or "").strip()
-    if not s:
-        return None
-    try:
-        val = json.loads(s)
-    except Exception:
-        return None
-    return val if isinstance(val, dict) else None
-
-
-def validate_skill_io_schema(raw: Any, *, field: str) -> tuple[dict[str, Any] | None, list[str]]:
-    """Strong-ish check: must be a JSON object (optional draft-like constraints)."""
-    errs: list[str] = []
-    if raw is None or raw == "" or raw is False:
-        return None, errs
-    obj = _parse_json_object(raw)
-    if obj is None:
-        errs.append(f"{field}_invalid_json_object")
-        return None, errs
-    t = obj.get("type")
-    if t is not None and str(t).lower() not in ("object", "array", "string", "number", "boolean"):
-        errs.append(f"{field}_unsupported_type:{t}")
-    if "properties" in obj and not isinstance(obj.get("properties"), dict):
-        errs.append(f"{field}_properties_must_be_object")
-    if "required" in obj and not isinstance(obj.get("required"), list):
-        errs.append(f"{field}_required_must_be_list")
-    if "allowed_ops" in obj and not isinstance(obj.get("allowed_ops"), list):
-        errs.append(f"{field}_allowed_ops_must_be_list")
-    return obj, errs
-
-
-def validate_against_schema(schema: dict[str, Any] | None, data: Any) -> list[str]:
-    """Minimal JSON-Schema-like validator for skill input args / output contracts."""
-    if not schema:
-        return []
-    errs: list[str] = []
-    st = str(schema.get("type") or "object").lower()
-    if st == "object":
-        if data is None:
-            data = {}
-        if not isinstance(data, dict):
-            return [f"expected_object_got_{type(data).__name__}"]
-        props = schema.get("properties") if isinstance(schema.get("properties"), dict) else {}
-        required = schema.get("required") if isinstance(schema.get("required"), list) else []
-        for key in required:
-            k = str(key)
-            if k not in data:
-                errs.append(f"missing_required:{k}")
-        for key, sub in props.items():
-            if key not in data or not isinstance(sub, dict):
-                continue
-            errs.extend(
-                f"{key}.{e}" for e in validate_against_schema(sub, data.get(key))
-            )
-        return errs
-    if st == "array":
-        if not isinstance(data, list):
-            return ["expected_array"]
-        try:
-            mn = int(schema["minItems"]) if "minItems" in schema else None
-            mx = int(schema["maxItems"]) if "maxItems" in schema else None
-        except (TypeError, ValueError):
-            mn = mx = None
-        if mn is not None and len(data) < mn:
-            errs.append(f"minItems:{mn}")
-        if mx is not None and len(data) > mx:
-            errs.append(f"maxItems:{mx}")
-        return errs
-    if st == "string" and not isinstance(data, str):
-        return ["expected_string"]
-    if st == "number" and not isinstance(data, (int, float)):
-        return ["expected_number"]
-    if st == "boolean" and not isinstance(data, bool):
-        return ["expected_boolean"]
-    if "enum" in schema and isinstance(schema.get("enum"), list):
-        if data not in schema["enum"]:
-            errs.append("enum_mismatch")
-    return errs
-
-
-def validate_skill_meta(item: dict[str, Any], *, source: str) -> list[str]:
-    """Validate seed/file/admin skill registration payload before upsert."""
-    errs: list[str] = []
-    key = str(item.get("skill_key") or item.get("skillKey") or "").strip()
-    if not key:
-        errs.append("skill_key_required")
-    name = str(item.get("name") or "").strip()
-    if not name:
-        errs.append("name_required")
-    body = str(item.get("prompt_positive") or item.get("promptPositive") or "").strip()
-    if not body:
-        errs.append("prompt_positive_required")
-    ns = _normalize_namespace(
-        item.get("namespace"), source=source
-    )
-    ns_prefix, local = split_namespace_key(key)
-    if ns_prefix and ns_prefix != ns:
-        errs.append(f"namespace_key_mismatch:{ns_prefix}!={ns}")
-    local_key = local or key
-    if source == SOURCE_ADMIN:
-        if local_key in _CORE_RESERVED_KEYS or (
-            ns_prefix == NS_CORE or (not ns_prefix and local_key in _CORE_RESERVED_KEYS)
-        ):
-            errs.append(f"core_key_reserved:{local_key}")
-        if ns_prefix == NS_EXT:
-            errs.append("user_skill_cannot_use_ext_namespace")
-    if source == SOURCE_FILE and local_key in _CORE_RESERVED_KEYS:
-        errs.append(f"file_pack_collides_core:{local_key}")
-    prefs = item.get("preferred_tools") or item.get("preferredTools") or []
-    if prefs is not None and not isinstance(prefs, (list, str)):
-        errs.append("preferred_tools_invalid")
-    _, in_errs = validate_skill_io_schema(
-        item.get("input_schema") or item.get("inputSchema"), field="input_schema"
-    )
-    errs.extend(in_errs)
-    _, out_errs = validate_skill_io_schema(
-        item.get("output_schema") or item.get("outputSchema"), field="output_schema"
-    )
-    errs.extend(out_errs)
-    ar = item.get("allowed_resources") or item.get("allowedResources")
-    if ar is not None:
-        parsed = _parse_allowed_resources(ar)
-        if parsed is None and not isinstance(ar, list) and str(ar).strip():
-            # empty list is valid; unparsable non-empty → error
-            if _parse_allowed_resources(ar) is None and str(ar).strip() not in ("[]",):
-                try:
-                    json.loads(str(ar))
-                except Exception:
-                    errs.append("allowed_resources_invalid")
-    return errs
-
-
-def runtime_skill_keys() -> frozenset[str]:
-    global _RUNTIME_SKILL_KEYS
-    if _RUNTIME_SKILL_KEYS is not None:
-        return _RUNTIME_SKILL_KEYS
-    keys = {str(k).strip() for k in _SEED_BY_KEY if str(k).strip()}
-    # Qualified aliases for core.
-    for k in list(keys):
-        keys.add(f"{NS_CORE}.{k}")
-    try:
-        with Session(core_db.engine) as session:
-            rows = crud.list_design_skill_keys_enabled(session=session)
-            for k, ns_raw in rows:
-                k = str(k or "").strip()
-                if not k:
-                    continue
-                keys.add(k)
-                ns = _normalize_namespace(ns_raw, source=None)
-                # If row has bare key, also accept qualified form.
-                ns_prefix, local = split_namespace_key(k)
-                if not ns_prefix:
-                    keys.add(f"{ns}.{k}")
-                    if ns == NS_CORE:
-                        keys.add(k)
-                else:
-                    keys.add(local)
-                    keys.add(k)
-    except Exception:
-        pass
-    _RUNTIME_SKILL_KEYS = frozenset(keys)
-    return _RUNTIME_SKILL_KEYS
-
-
-def invalidate_skill_key_cache() -> None:
-    global _RUNTIME_SKILL_KEYS, _RUNTIME_SKILL_INDEX
-    _RUNTIME_SKILL_KEYS = None
-    _RUNTIME_SKILL_INDEX = None
-
-
-def reset_skills_ready_for_tests() -> None:
-    """Test helper: force ensure_design_skills to run again."""
-    global _SKILLS_READY, _RUNTIME_SKILL_KEYS, _RUNTIME_SKILL_INDEX, _DISK_SIGNATURE
-    stop_skills_hot_reload()
-    _SKILLS_READY = False
-    _RUNTIME_SKILL_KEYS = None
-    _RUNTIME_SKILL_INDEX = None
-    _DISK_SIGNATURE = None
-
-
-def _csv_has(csv: str, token: str) -> bool:
-    parts = {p.strip().lower() for p in str(csv or "").split(",") if p.strip()}
-    if not parts or "all" in parts:
-        return True
-    return token.strip().lower() in parts
-
-
-def _row_get(r: Any, key: str, default: Any = None) -> Any:
-    try:
-        if hasattr(r, "keys") and key in r.keys():
-            return r[key]
-    except Exception:
-        pass
-    if isinstance(r, dict):
-        return r.get(key, default)
-    if hasattr(r, key):
-        try:
-            return getattr(r, key)
-        except Exception:
-            return default
-    return default
-
-
-def _parse_preferred_tools(raw: Any) -> list[str]:
-    if isinstance(raw, list):
-        return [str(x).strip() for x in raw if str(x).strip()][:24]
-    s = str(raw or "").strip()
-    if not s:
-        return []
-    try:
-        val = json.loads(s)
-        if isinstance(val, list):
-            return [str(x).strip() for x in val if str(x).strip()][:24]
-    except Exception:
-        pass
-    return [p.strip() for p in s.replace("；", ",").split(",") if p.strip()][:24]
-
-
-_INTERNAL_RESOURCE_KINDS = frozenset({"knowledge", "aesthetics", "tools"})
-
-
-def _parse_allowed_resources(raw: Any) -> list[str] | None:
-    """None = unspecified; list may be empty (= deny all internal resources)."""
-    if raw is None:
-        return None
-    if isinstance(raw, list):
-        out = []
-        for x in raw:
-            k = str(x or "").strip().lower()
-            if k in _INTERNAL_RESOURCE_KINDS and k not in out:
-                out.append(k)
-        return out
-    s = str(raw or "").strip()
-    if not s:
-        return []
-    try:
-        val = json.loads(s)
-        if isinstance(val, list):
-            return _parse_allowed_resources(val)
-    except Exception:
-        pass
-    parts = [p.strip().lower() for p in s.replace("；", ",").split(",") if p.strip()]
-    return [p for p in parts if p in _INTERNAL_RESOURCE_KINDS]
-
-
-def _parse_triggers(raw: Any) -> list[dict[str, Any]]:
-    if raw is None or raw is False:
-        return []
-    if isinstance(raw, dict):
-        return [raw]
-    if isinstance(raw, list):
-        return [x for x in raw if isinstance(x, dict)]
-    s = str(raw or "").strip()
-    if not s:
-        return []
-    try:
-        val = json.loads(s)
-    except Exception:
-        return []
-    if isinstance(val, dict):
-        return [val]
-    if isinstance(val, list):
-        return [x for x in val if isinstance(x, dict)]
-    return []
-
-
-def _normalize_source(raw: Any, *, default: str = SOURCE_ADMIN) -> str:
-    s = str(raw or "").strip().lower()
-    if s in (SOURCE_SEED, SOURCE_ADMIN, SOURCE_FILE):
-        return s
-    return default
-
-
-def _parse_pack_version(raw: Any) -> tuple[str, int]:
-    """Return (label, sortable int). Accepts 1 / '1' / '1.0.0'."""
-    s = str(raw if raw is not None else "1").strip() or "1"
-    if re.fullmatch(r"-?\d+", s):
-        return s, max(1, int(s))
-    m = re.match(r"^(\d+)(?:\.(\d+))?(?:\.(\d+))?", s)
-    if m:
-        major = int(m.group(1))
-        minor = int(m.group(2) or 0)
-        patch = int(m.group(3) or 0)
-        return s, max(1, major * 1_000_000 + minor * 1_000 + patch)
-    return s, 1
-
-
-def _read_json_file(path: Path) -> dict[str, Any] | None:
-    try:
-        data = json.loads(path.read_text(encoding="utf-8"))
-    except Exception:
-        return None
-    return data if isinstance(data, dict) else None
-
-
-def _locale_pick(
-    locales: Any, *, prefer: tuple[str, ...] = ("zh-CN", "zh", "en-US", "en")
-) -> dict[str, Any]:
-    if not isinstance(locales, dict):
-        return {}
-    for key in prefer:
-        block = locales.get(key)
-        if isinstance(block, dict):
-            return block
-    for block in locales.values():
-        if isinstance(block, dict):
-            return block
-    return {}
-
-
-def _resolve_pack_logo(pack_dir: Path, meta: dict[str, Any]) -> str:
-    """Return path relative to design_skills root, or absolute URL, or ''."""
-    raw = str(meta.get("logo") or "").strip()
-    if raw.startswith(("http://", "https://", "data:")):
-        return raw
-    candidates: list[Path] = []
-    if raw:
-        p = Path(raw)
-        # Never accept absolute paths outside the pack (path traversal / leakage).
-        if p.is_absolute():
-            return ""
-        candidates.append(pack_dir / p)
-        candidates.append(pack_dir / Path(raw).name)
-    key = pack_dir.name
-    candidates.extend(
-        [
-            pack_dir / f"{key}-logo.png",
-            pack_dir / f"{key}-logo.svg",
-            pack_dir / f"{key}-logo.webp",
-            pack_dir / f"{key}-logo.jpg",
-            pack_dir / "logo.png",
-            pack_dir / "logo.svg",
-            pack_dir / "logo.webp",
-            pack_dir / "logo.jpg",
-        ]
-    )
-    root = pack_dir.parent.resolve()
-    for cand in candidates:
-        try:
-            if not cand.is_file():
-                continue
-            resolved = cand.resolve()
-            try:
-                rel = resolved.relative_to(root)
-                return rel.as_posix()
-            except ValueError:
-                # Outside this pack's design_skills root — deny.
-                continue
-        except Exception:
-            continue
-    return ""
-
-
-def _skill_item_from_parts(
-    *,
-    pack_dir: Path,
-    meta: dict[str, Any],
-    body: str,
-    skill_md_path: Path,
-) -> dict[str, Any] | None:
-    """Build skill dict from _meta.json + SKILL.md body."""
-    folder = pack_dir.name
-    # _meta.json: `name` is technical id; displayName lives in locales.
-    key = str(meta.get("skill_key") or meta.get("skillKey") or meta.get("name") or folder).strip()
-    if not key or key in _SEED_BY_KEY:
-        return None
-    locales = meta.get("locales") if isinstance(meta.get("locales"), dict) else {}
-    loc = _locale_pick(locales)
-    display = str(
-        loc.get("displayName")
-        or loc.get("display_name")
-        or meta.get("displayName")
-        or meta.get("display_name")
-        or meta.get("title")
-        or key
-    ).strip() or key
-    description = str(
-        loc.get("description")
-        or meta.get("description")
-        or meta.get("when_to_use")
-        or meta.get("whenToUse")
-        or ""
-    ).strip()
-    when = str(meta.get("when_to_use") or meta.get("whenToUse") or description).strip()
-    pos = body.strip()
-    if not pos:
-        return None
-    pack_label, ver_int = _parse_pack_version(meta.get("version") or meta.get("pack_version") or 1)
-    logo = _resolve_pack_logo(pack_dir, meta)
-    ns_prefix, stripped = split_namespace_key(key)
-    if ns_prefix == NS_CORE:
-        return None
-    storage_key = qualify_skill_key(NS_EXT, stripped or key) if ns_prefix == NS_EXT else (
-        stripped or key
-    )
-    # Bare file-pack keys stay bare for BC; namespace column still marks ext.
-    if not ns_prefix:
-        storage_key = (stripped or key).strip().lower()
-    if storage_key in _SEED_BY_KEY or (stripped or key) in _SEED_BY_KEY:
-        return None
-    meta_errs = validate_skill_meta(
-        {
-            "skill_key": storage_key,
-            "name": display,
-            "prompt_positive": pos,
-            "preferred_tools": meta.get("preferred_tools") or meta.get("preferredTools") or [],
-            "allowed_resources": meta.get("allowed_resources")
-            or meta.get("allowedResources"),
-            "input_schema": meta.get("input_schema") or meta.get("inputSchema"),
-            "output_schema": meta.get("output_schema") or meta.get("outputSchema"),
-            "namespace": NS_EXT,
-        },
-        source=SOURCE_FILE,
-    )
-    if meta_errs:
-        logger.warning("skip skill pack %s: %s", pack_dir.name, ",".join(meta_errs))
-        return None
-    return {
-        "skill_key": storage_key,
-        "name": display,
-        "description": description,
-        "category": str(meta.get("category") or "agent").strip() or "agent",
-        "when_to_use": when,
-        "prompt_positive": pos,
-        "prompt_negative": str(meta.get("prompt_negative") or meta.get("promptNegative") or "").strip(),
-        "scenes": str(meta.get("scenes") or "all").strip() or "all",
-        "sort_weight": int(meta.get("sort_weight") or meta.get("sortWeight") or 0),
-        "preferred_tools": meta.get("preferred_tools") or meta.get("preferredTools") or [],
-        "allowed_resources": meta.get("allowed_resources")
-        or meta.get("allowedResources"),
-        "input_schema": meta.get("input_schema") or meta.get("inputSchema"),
-        "output_schema": meta.get("output_schema") or meta.get("outputSchema"),
-        "triggers": meta.get("triggers") or [],
-        "mutex_group": str(meta.get("mutex_group") or meta.get("mutexGroup") or "").strip(),
-        "version": ver_int,
-        "pack_version": pack_label,
-        "logo": logo,
-        "locales": locales if isinstance(locales, dict) else {},
-        "source": SOURCE_FILE,
-        "namespace": NS_EXT,
-        "_path": str(skill_md_path),
-        "_pack": str(pack_dir),
-    }
-
-
-def _load_pack_dir(pack_dir: Path) -> dict[str, Any] | None:
-    """Load one skill pack: ``_meta.json`` + ``SKILL.md``, or frontmatter-only ``SKILL.md``."""
-    if not pack_dir.is_dir():
-        return None
-    skill_md = _skill_md_path(pack_dir)
-    if not skill_md:
-        return None
-    try:
-        raw = skill_md.read_text(encoding="utf-8")
-    except Exception:
-        return None
-    fm, body = _split_skill_md_frontmatter(raw)
-    if not body:
-        return None
-    meta: dict[str, Any] | None = None
-    for name in _META_NAMES:
-        p = pack_dir / name
-        if p.is_file():
-            meta = _read_json_file(p)
-            if meta:
-                break
-    if not meta:
-        meta = _meta_from_agent_skill_frontmatter(fm, folder=pack_dir.name)
-    if not meta:
-        return None
-    return _skill_item_from_parts(
-        pack_dir=pack_dir,
-        meta=meta,
-        body=body,
-        skill_md_path=skill_md,
-    )
-
-
-def _load_file_skills() -> list[dict[str, Any]]:
-    """Scan ``.agents/skills`` + design_skills dirs → skill dicts (later roots win).
-
-    Under ``.agents/skills``, only packs with product ``_meta.json`` are ingested
-    (skips IDE-only docs such as ``langfuse``).
-    """
-    by_key: dict[str, dict[str, Any]] = {}
-    agents_root = _agents_skills_dir()
-    try:
-        agents_resolved = agents_root.resolve()
-    except Exception:
-        agents_resolved = agents_root
-    for root in _file_skills_dirs():
-        try:
-            root_resolved = root.resolve()
-        except Exception:
-            root_resolved = root
-        require_product_meta = root_resolved == agents_resolved
-        for pack_dir in sorted(p for p in root.iterdir() if p.is_dir()):
-            if require_product_meta and not _pack_has_product_meta(pack_dir):
-                continue
-            item = _load_pack_dir(pack_dir)
-            if not item:
-                continue
-            key = str(item.get("skill_key") or "").strip()
-            if not key:
-                continue
-            by_key[key] = item
-    return list(by_key.values())
-
-
-def _pub(r: Any) -> dict[str, Any]:
-    key = str(_row_get(r, "skill_key") or "").strip()
-    preferred = _parse_preferred_tools(_row_get(r, "preferred_tools"))
-    triggers = _parse_triggers(_row_get(r, "triggers"))
-    source = _normalize_source(_row_get(r, "source"), default=SOURCE_ADMIN)
-    namespace = _normalize_namespace(_row_get(r, "namespace"), source=source)
-    allowed_resources = _parse_allowed_resources(_row_get(r, "allowed_resources"))
-    input_schema = _parse_json_object(_row_get(r, "input_schema"))
-    output_schema = _parse_json_object(_row_get(r, "output_schema"))
-    locales_raw = _row_get(r, "locales")
-    locales: dict[str, Any] = {}
-    if isinstance(locales_raw, dict):
-        locales = locales_raw
-    elif locales_raw:
-        try:
-            val = json.loads(str(locales_raw))
-            if isinstance(val, dict):
-                locales = val
-        except Exception:
-            locales = {}
-    ns_prefix, local = split_namespace_key(key)
-    qualified = key if ns_prefix else f"{namespace}.{key}" if key else ""
-    return {
-        "id": int(_row_get(r, "id") or 0),
-        "skillKey": key or None,
-        "qualifiedKey": qualified or None,
-        "name": str(_row_get(r, "name") or ""),
-        "description": str(_row_get(r, "description") or ""),
-        "category": str(_row_get(r, "category") or ""),
-        "whenToUse": str(_row_get(r, "when_to_use") or ""),
-        "promptPositive": str(_row_get(r, "prompt_positive") or ""),
-        "promptNegative": str(_row_get(r, "prompt_negative") or ""),
-        "preferredTools": preferred,
-        "allowedResources": allowed_resources,
-        "inputSchema": input_schema,
-        "outputSchema": output_schema,
-        "triggers": triggers,
-        "mutexGroup": str(_row_get(r, "mutex_group") or "").strip() or None,
-        "version": int(_row_get(r, "version") or 1),
-        "packVersion": str(_row_get(r, "pack_version") or "").strip() or None,
-        "logo": str(_row_get(r, "logo") or "").strip() or None,
-        "locales": locales,
-        "source": source,
-        "namespace": namespace,
-        "skillKind": skill_kind_for_namespace(namespace),
-        "ownerUserId": str(_row_get(r, "owner_user_id") or "").strip() or None,
-        "sortWeight": int(_row_get(r, "sort_weight") or 0),
-        "scenes": str(_row_get(r, "scenes") or "all"),
-        "enabled": bool(int(_row_get(r, "enabled") or 0)),
-        "_localKey": local or key,
-    }
-
-
-def _load_user_skill_prefs(user_id: str) -> dict[str, bool]:
-    """skill_key(lower) → enabled for this user (missing = default on)."""
-    uid = str(user_id or "").strip()
-    if not uid:
-        return {}
-    ensure_design_skills()
-    with Session(core_db.engine) as session:
-        rows = crud.list_user_skill_prefs(session=session, user_id=uid)
-    out: dict[str, bool] = {}
-    for r in rows:
-        key = str(_row_get(r, "skill_key") or "").strip().lower()
-        if not key:
-            continue
-        out[key] = bool(int(_row_get(r, "enabled") or 0))
-    return out
-
-
-def list_runtime_skills(
-    *,
-    scene: str = "website",
-    enabled_only: bool = True,
-    user_id: str | None = None,
-) -> list[dict[str, Any]]:
-    ensure_design_skills()
-    scene_l = str(scene or "website").strip().lower() or "website"
-    uid = str(user_id or "").strip() or None
-    prefs = _load_user_skill_prefs(uid) if uid else {}
-    with Session(core_db.engine) as session:
-        rows = crud.list_design_skills_runtime(
-            session=session, enabled_only=enabled_only
-        )
-    out: list[dict[str, Any]] = []
-    seen: set[str] = set()
-    for r in rows:
-        item = _pub(r)
-        key = str(item.get("skillKey") or "").strip()
-        if not key or key in seen:
-            continue
-        ns = str(item.get("namespace") or NS_USER)
-        owner = str(item.get("ownerUserId") or "").strip() or None
-        # User-owned extension skills are isolated to that user.
-        if ns == NS_USER and owner and uid and owner != uid:
-            continue
-        if ns == NS_USER and owner and not uid:
-            continue
-        if not (
-            _csv_has(item.get("scenes") or "all", scene_l)
-            or _csv_has(item.get("scenes") or "all", "all")
-        ):
-            continue
-        # Per-user off switch (official + others); missing pref = on.
-        if enabled_only and prefs and prefs.get(key.lower()) is False:
-            continue
-        seen.add(key)
-        out.append(item)
-    return out
-
-
-def resolve_storage_skill_key(raw: str, *, scene: str = "website") -> str | None:
-    """Map bare / namespaced / legacy refs onto the storage skill_key."""
-    base, _, _ = parse_skill_pin(raw)
-    s = base.strip().lower()
-    if not s:
-        return None
-    if s.startswith("skill."):
-        s = s[6:]
-    if s.startswith("skill_"):
-        s = s[6:]
-    ns, local = split_namespace_key(s)
-    rows = list_runtime_skills(scene=scene, enabled_only=True)
-    by_key = {str(r.get("skillKey") or "").strip().lower(): r for r in rows}
-
-    def _match(want_ns: str | None, want_local: str) -> str | None:
-        # Exact storage key
-        if want_local in by_key and (
-            want_ns is None or str(by_key[want_local].get("namespace")) == want_ns
-        ):
-            return want_local
-        qualified = f"{want_ns}.{want_local}" if want_ns else ""
-        if qualified and qualified in by_key:
-            return qualified
-        for k, row in by_key.items():
-            row_ns = str(row.get("namespace") or "")
-            row_local = str(row.get("_localKey") or k).lower()
-            if want_ns and row_ns != want_ns:
-                continue
-            if row_local == want_local or k == want_local or k.endswith(f".{want_local}"):
-                return k
-        return None
-
-    if ns:
-        hit = _match(ns, local)
-        if hit:
-            return hit
-        return None
-    # Bare: prefer core → ext → user
-    for prefer in (NS_CORE, NS_EXT, NS_USER):
-        hit = _match(prefer, local)
-        if hit:
-            return hit
-    if local in by_key:
-        return local
-    return None
-
-
-def _apply_mutex(rows: list[dict[str, Any]]) -> list[dict[str, Any]]:
-    """Keep highest sortWeight per mutex_group (rows already weight-desc)."""
-    seen_g: set[str] = set()
-    out: list[dict[str, Any]] = []
-    for r in rows:
-        g = str(r.get("mutexGroup") or "").strip().lower()
-        if g:
-            if g in seen_g:
-                continue
-            seen_g.add(g)
-        out.append(r)
-    return out
-
-
-def format_skills_catalog(*, scene: str = "website", user_id: str | None = None) -> str:
-    rows = list_runtime_skills(scene=scene, user_id=user_id)
-    from app.services.design.prompts.prompt_pack_store import render_prompt_body
-
-    header = render_prompt_body("agent.prompt.skill_catalog_header").strip()
-    lines = [header] if header else []
-    for r in rows:
-        key = str(r.get("skillKey") or "").strip()
-        if not key:
-            continue
-        name = str(r.get("name") or key).strip()
-        when = str(r.get("whenToUse") or "").strip()
-        ver = int(r.get("version") or 1)
-        ns = str(r.get("namespace") or NS_USER)
-        kind = str(r.get("skillKind") or skill_kind_for_namespace(ns))
-        q = str(r.get("qualifiedKey") or key)
-        line = f"- `{key}` [{ns}/{kind}] v{ver}"
-        if q != key:
-            line += f" alias:`{q}`"
-        line += f" — {name}"
-        if when:
-            line += f"（{when[:80]}）"
-        lines.append(line)
-        if len(lines) >= 16:
-            break
-    if len(lines) <= (1 if header else 0):
-        empty = render_prompt_body("agent.prompt.skill_catalog_empty").strip()
-        if empty:
-            lines.append(empty)
-    return "\n".join(lines)
-
-
-def _load_skill_revision_snapshot(
-    *,
-    skill_key: str,
-    version: int | None = None,
-    pack_version: str | None = None,
-) -> dict[str, Any] | None:
-    key = str(skill_key or "").strip()
-    if not key:
-        return None
-    try:
-        with Session(core_db.engine) as session:
-            raw = crud.get_design_skill_revision_snapshot(
-                session=session,
-                skill_key=key,
-                version=version,
-                pack_version=pack_version,
-            )
-        if not raw:
-            return None
-        snap = _parse_json_object(raw)
-        return snap
-    except Exception:
-        return None
-
-
-def save_skill_revision(session: Session, *, skill_id: int, item: dict[str, Any]) -> None:
-    """Persist a version snapshot (best-effort; never breaks upsert)."""
-    try:
-        key = str(item.get("skillKey") or item.get("skill_key") or "").strip()
-        if not key:
-            return
-        ver = int(item.get("version") or 1)
-        snap = {
-            k: item.get(k)
-            for k in (
-                "skillKey",
-                "name",
-                "description",
-                "category",
-                "whenToUse",
-                "promptPositive",
-                "promptNegative",
-                "preferredTools",
-                "allowedResources",
-                "inputSchema",
-                "outputSchema",
-                "triggers",
-                "mutexGroup",
-                "version",
-                "packVersion",
-                "namespace",
-                "source",
-                "scenes",
-                "sortWeight",
-            )
-        }
-        crud.insert_design_skill_revision(
-            session=session,
-            skill_id=int(skill_id),
-            skill_key=key,
-            namespace=str(item.get("namespace") or NS_USER),
-            version=ver,
-            pack_version=str(item.get("packVersion") or item.get("pack_version") or "")
-            or None,
-            snapshot=json.dumps(snap, ensure_ascii=False),
-            source=str(item.get("source") or SOURCE_ADMIN),
-            created_at=time.time(),
-            commit=False,
-        )
-    except Exception:
-        logger.debug("skill revision save failed", exc_info=True)
-
-
-def format_skills_details(
-    *,
-    keys: list[str],
-    scene: str = "website",
-    max_chars: int = MAX_SKILL_DETAIL_CHARS,
-    user_id: str | None = None,
-    version_pins: dict[str, int | str] | None = None,
-    input_args: dict[str, Any] | None = None,
-) -> str:
-    text, _errs = format_skills_details_checked(
-        keys=keys,
-        scene=scene,
-        max_chars=max_chars,
-        user_id=user_id,
-        version_pins=version_pins,
-        input_args=input_args,
-    )
-    return text
-
-
-def format_skills_details_checked(
-    *,
-    keys: list[str],
-    scene: str = "website",
-    max_chars: int = MAX_SKILL_DETAIL_CHARS,
-    user_id: str | None = None,
-    version_pins: dict[str, int | str] | None = None,
-    input_args: dict[str, Any] | None = None,
-) -> tuple[str, list[str]]:
-    """Return (details_markdown, validation_errors)."""
-    wanted_raw = [str(k).strip() for k in (keys or []) if str(k).strip()]
-    if not wanted_raw:
-        return "", []
-    load_all = any(k in ("*", "all") for k in wanted_raw)
-    pins = version_pins or {}
-    args_by_key = input_args or {}
-    errs: list[str] = []
-    resolved: list[dict[str, Any]] = []
-    runtime = list_runtime_skills(scene=scene, user_id=user_id)
-    by_key = {str(r.get("skillKey") or "").strip().lower(): r for r in runtime}
-
-    if load_all:
-        resolved = list(runtime)
-    else:
-        for raw in wanted_raw:
-            base, pin_i, pin_p = parse_skill_pin(raw)
-            storage = resolve_storage_skill_key(base, scene=scene)
-            if not storage:
-                errs.append(f"skill_unknown:{raw}")
-                continue
-            row = dict(by_key.get(storage) or {})
-            if not row:
-                errs.append(f"skill_unavailable:{storage}")
-                continue
-            pin = pins.get(storage, pins.get(base))
-            if pin_i is not None:
-                pin = pin_i
-            elif pin_p is not None:
-                pin = pin_p
-            if pin is not None:
-                snap = _load_skill_revision_snapshot(
-                    skill_key=storage,
-                    version=int(pin) if isinstance(pin, int) or str(pin).isdigit() else None,
-                    pack_version=None if isinstance(pin, int) or str(pin).isdigit() else str(pin),
-                )
-                if snap:
-                    row = {**row, **snap, "skillKey": storage}
-                else:
-                    # Pin miss: refuse silent upgrade
-                    cur = int(row.get("version") or 0)
-                    if isinstance(pin, int) or str(pin).isdigit():
-                        if int(pin) != cur:
-                            errs.append(f"skill_version_missing:{storage}@{pin}")
-                            continue
-                    elif str(row.get("packVersion") or "") != str(pin):
-                        errs.append(f"skill_pack_version_missing:{storage}@{pin}")
-                        continue
-            schema = row.get("inputSchema") if isinstance(row.get("inputSchema"), dict) else None
-            if schema:
-                arg_errs = validate_against_schema(schema, args_by_key.get(storage) or {})
-                if arg_errs:
-                    errs.append(f"skill_input_invalid:{storage}:" + ",".join(arg_errs[:6]))
-                    continue
-            resolved.append(row)
-
-    rows = _apply_mutex(resolved)
-    from app.services.design.prompts.prompt_pack_store import render_prompt_body
-
-    header = render_prompt_body("agent.prompt.skill_details_header").strip()
-    parts: list[str] = [header] if header else []
-    total = len(parts[0]) if parts else 0
-    used = 0
-    for r in rows:
-        key = str(r.get("skillKey") or "").strip().lower()
-        if not key:
-            continue
-        name = str(r.get("name") or key)
-        when = str(r.get("whenToUse") or "").strip()
-        body = str(r.get("promptPositive") or "").strip()
-        neg = str(r.get("promptNegative") or "").strip()
-        tools = r.get("preferredTools") or []
-        ver = int(r.get("version") or 1)
-        ns = str(r.get("namespace") or NS_USER)
-        head = f"## skill: {key} — {name} (v{ver}, ns={ns})"
-        if when:
-            head += f"\nwhen: {when}"
-        if tools:
-            head += "\npreferred_tools: " + ", ".join(str(t) for t in tools)
-        out_schema = r.get("outputSchema") if isinstance(r.get("outputSchema"), dict) else None
-        if out_schema and isinstance(out_schema.get("allowed_ops"), list):
-            head += "\noutput_allowed_ops: " + ", ".join(
-                str(t) for t in out_schema.get("allowed_ops") if str(t).strip()
-            )
-        block = f"{head}\n{body}".strip()
-        if neg:
-            block += f"\n\nforbid: {neg}"
-        if total + len(block) + 2 > max_chars and used > 0:
-            trunc = render_prompt_body("agent.prompt.skill_details_truncated").strip()
-            if trunc:
-                parts.append(trunc)
-            break
-        parts.append(block)
-        total += len(block) + 2
-        used += 1
-        if load_all and used >= 12:
-            break
-    if used == 0:
-        return "", errs
-    return "\n\n".join(parts), errs
-
-
-def _normalize_loaded_skill_keys(
-    skill_keys: list[str], *, scene: str = "website"
-) -> set[str]:
-    """Map need_skills refs (bare / namespaced / aliases) → storage skill_keys."""
-    raw = {str(k).strip().lower() for k in skill_keys if str(k).strip()}
-    if not raw:
-        return set()
-    out: set[str] = set()
-    for k in raw:
-        hit = resolve_storage_skill_key(k, scene=scene)
-        out.add((hit or k).lower())
-    return out
-
-
-def _iter_skills_for_keys(
-    skill_keys: list[str], *, scene: str = "website"
-):
-    keys = _normalize_loaded_skill_keys(skill_keys, scene=scene)
-    if not keys:
-        return
-    for r in list_runtime_skills(scene=scene):
-        key = str(r.get("skillKey") or "").strip().lower()
-        if key in keys:
-            yield r
-
-
-def preferred_tools_allowlist(
-    skill_keys: list[str], *, scene: str = "website"
-) -> set[str] | None:
-    """Union of preferred_tools for loaded skills.
-
-    None = no restriction (only platform seed/file skills with empty prefs).
-    Custom admin skills without preferred_tools → hard-restrict to layout ops.
-    """
-    if not _normalize_loaded_skill_keys(skill_keys, scene=scene):
-        return None
-    allow: set[str] = set()
-    any_pref = False
-    any_custom_unscoped = False
-    only_custom = True
-    for r in _iter_skills_for_keys(skill_keys, scene=scene):
-        source = str(r.get("source") or "").strip().lower()
-        ns = str(r.get("namespace") or _SOURCE_TO_NS.get(source, NS_USER))
-        if source in (SOURCE_SEED, SOURCE_FILE) or ns in (NS_CORE, NS_EXT):
-            only_custom = False
-        prefs = r.get("preferredTools") or []
-        if prefs:
-            any_pref = True
-            allow.update(str(t).strip() for t in prefs if str(t).strip())
-        elif source == SOURCE_ADMIN or ns == NS_USER:
-            any_custom_unscoped = True
-    if any_pref:
-        allow |= _ALWAYS_ALLOW_OPS
-        return allow
-    if only_custom or any_custom_unscoped:
-        # Custom skills must not inherit unrestricted canvas tool surface.
-        return set(_ALWAYS_ALLOW_OPS)
-    return None
-
-
-def skill_resource_allowlist(
-    skill_keys: list[str], *, scene: str = "website"
-) -> set[str] | None:
-    """Which internal need_* resources loaded skills may unlock.
-
-    None = unrestricted (platform-only skills without explicit ACL).
-    set() = deny knowledge / prompts / aesthetics.
-    Any user-extension skill in the load set → never unrestricted.
-    """
-    if not _normalize_loaded_skill_keys(skill_keys, scene=scene):
-        return None
-    allowed: set[str] = set()
-    saw_custom = False
-    saw_platform_open = False
-    for r in _iter_skills_for_keys(skill_keys, scene=scene):
-        source = str(r.get("source") or "").strip().lower()
-        ns = str(r.get("namespace") or _SOURCE_TO_NS.get(source, NS_USER))
-        res = r.get("allowedResources")
-        if isinstance(res, list):
-            allowed.update(str(x).strip().lower() for x in res if str(x).strip())
-            if source == SOURCE_ADMIN or ns == NS_USER:
-                saw_custom = True
-            continue
-        if source in (SOURCE_SEED, SOURCE_FILE) or ns in (NS_CORE, NS_EXT):
-            saw_platform_open = True
-        elif source == SOURCE_ADMIN or ns == NS_USER:
-            saw_custom = True
-    # Hard isolation: user-extension skills never inherit unrestricted surface.
-    if saw_custom:
-        if not allowed:
-            return {"tools"}
-        return allowed | {"tools"}
-    if saw_platform_open:
-        return None
-    return None
-
-
-def filter_ops_by_skill_output_schema(
-    ops: list[dict[str, Any]],
-    *,
-    skill_keys: list[str],
-    scene: str = "website",
-) -> tuple[list[dict[str, Any]], list[str]]:
-    """Enforce union of output_schema.allowed_ops when declared by loaded skills."""
-    if not _normalize_loaded_skill_keys(skill_keys, scene=scene):
-        return list(ops or []), []
-    allowed_ops: set[str] = set()
-    any_schema = False
-    for r in _iter_skills_for_keys(skill_keys, scene=scene):
-        schema = r.get("outputSchema") if isinstance(r.get("outputSchema"), dict) else None
-        if not schema:
-            continue
-        ops_list = schema.get("allowed_ops")
-        if isinstance(ops_list, list) and ops_list:
-            any_schema = True
-            allowed_ops.update(str(x).strip() for x in ops_list if str(x).strip())
-    if not any_schema:
-        return list(ops or []), []
-    allowed_ops |= _ALWAYS_ALLOW_OPS
-    kept: list[dict[str, Any]] = []
-    from app.services.design.ops.tool_ops_contract import format_op_error
-
-    errs: list[str] = []
-    for op in ops or []:
-        if not isinstance(op, dict):
-            continue
-        name = str(op.get("name") or op.get("op_key") or "").strip()
-        if not name or name in allowed_ops:
-            kept.append(op)
-            continue
-        errs.append(
-            format_op_error(
-                "op_not_in_skill_output_schema",
-                fix="omit this op or load a skill that allows it",
-                detail=f"name={name}",
-            )
-        )
-    return kept, errs
-
-
-def filter_need_resources_by_skill_acl(
-    *,
-    skill_keys: list[str],
-    scene: str,
-    need_knowledge: list[str],
-    need_aesthetics: bool,
-) -> tuple[list[str], bool, list[str]]:
-    """Drop internal resource requests blocked by custom-skill ACL."""
-    allow = skill_resource_allowlist(skill_keys, scene=scene)
-    if allow is None:
-        return list(need_knowledge or []), bool(need_aesthetics), []
-    errs: list[str] = []
-    k = list(need_knowledge or [])
-    a = bool(need_aesthetics)
-    if k and "knowledge" not in allow:
-        errs.append("skill_acl_deny:knowledge")
-        k = []
-    if a and "aesthetics" not in allow:
-        errs.append("skill_acl_deny:aesthetics")
-        a = False
-    return k, a, errs
-
-
-def filter_ops_by_skill_allowlist(
-    ops: list[dict[str, Any]],
-    *,
-    skill_keys: list[str],
-    scene: str = "website",
-) -> tuple[list[dict[str, Any]], list[str]]:
-    """Enforce preferred_tools / custom-skill hard allowlist + output_schema."""
-    from app.services.design.ops.tool_ops_contract import format_op_error
-
-    allow = preferred_tools_allowlist(skill_keys, scene=scene)
-    kept: list[dict[str, Any]] = []
-    errs: list[str] = []
-    if allow is None:
-        kept = list(ops or [])
-    else:
-        for op in ops or []:
-            if not isinstance(op, dict):
-                continue
-            name = str(op.get("name") or op.get("op_key") or "").strip()
-            if not name or name in allow:
-                kept.append(op)
-                continue
-            errs.append(
-                format_op_error(
-                    "op_not_in_skill_allowlist",
-                    fix="omit this op or use a preferred_tools skill",
-                    detail=f"name={name}",
-                )
-            )
-    kept2, errs2 = filter_ops_by_skill_output_schema(kept, skill_keys=skill_keys, scene=scene)
-    return kept2, errs + errs2
-
-
-def parse_need_skills_with_pins(
-    raw: Any, *, max_n: int = 8, scene: str = "website"
-) -> tuple[list[str], dict[str, int | str], dict[str, Any], list[str]]:
-    """Parse need_skills → (storage_keys, version_pins, input_args, errors)."""
-    errs: list[str] = []
-    if raw is None or raw is False:
-        return [], {}, {}, []
-    if raw is True:
-        return ["*"], {}, {}, []
-    items: list[Any]
-    if isinstance(raw, str):
-        s = raw.strip()
-        if s.lower() in ("1", "true", "yes", "all", "*"):
-            return ["*"], {}, {}, []
-        items = [p.strip() for p in s.replace("；", ",").split(",")]
-    elif isinstance(raw, list):
-        items = raw
-    else:
-        return [], {}, {}, ["need_skills_invalid_type"]
-
-    out: list[str] = []
-    pins: dict[str, int | str] = {}
-    args: dict[str, Any] = {}
-    seen: set[str] = set()
-    known = runtime_skill_keys()
-
-    for item in items:
-        pin_i: int | None = None
-        pin_p: str | None = None
-        arg_obj: Any = None
-        if isinstance(item, dict):
-            key_raw = str(
-                item.get("key")
-                or item.get("skill_key")
-                or item.get("skillKey")
-                or item.get("id")
-                or ""
-            ).strip()
-            if "version" in item and item.get("version") is not None:
-                try:
-                    pin_i = int(item.get("version"))
-                except (TypeError, ValueError):
-                    pin_p = str(item.get("version"))
-            if item.get("packVersion") or item.get("pack_version"):
-                pin_p = str(item.get("packVersion") or item.get("pack_version"))
-            if "args" in item:
-                arg_obj = item.get("args")
-        else:
-            key_raw = str(item or "").strip()
-        if not key_raw:
-            continue
-        base, from_at_i, from_at_p = parse_skill_pin(key_raw)
-        if pin_i is None and from_at_i is not None:
-            pin_i = from_at_i
-        if pin_p is None and from_at_p is not None:
-            pin_p = from_at_p
-        key = base.strip().lower()
-        if not key or key in seen:
-            continue
-        if key in ("all", "*"):
-            return ["*"], {}, {}, errs
-        if key.startswith("skill."):
-            key = key[6:]
-        if key.startswith("skill_"):
-            key = key[6:]
-        storage = resolve_storage_skill_key(key, scene=scene) or key
-        if known and storage not in known and key not in known and f"core.{key}" not in known:
-            # Also accept qualified forms already in known
-            ns, local = split_namespace_key(key)
-            alt = f"{ns}.{local}" if ns else f"core.{key}"
-            if alt not in known and storage not in known:
-                errs.append(f"skill_unknown:{key_raw}")
-                continue
-            storage = resolve_storage_skill_key(key, scene=scene) or storage
-        if storage in seen:
-            continue
-        seen.add(storage)
-        out.append(storage)
-        if pin_i is not None:
-            pins[storage] = pin_i
-        elif pin_p is not None:
-            pins[storage] = pin_p
-        if arg_obj is not None:
-            args[storage] = arg_obj
-        if len(out) >= max_n:
-            break
-    return out, pins, args, errs
-
-
-def normalize_need_skills(raw: Any, *, max_n: int = 8) -> list[str]:
-    keys, _pins, _args, _errs = parse_need_skills_with_pins(raw, max_n=max_n)
-    return keys
-
-
-def _rule_matches(
-    rule: dict[str, Any],
-    *,
-    empty_canvas: bool,
-    has_images: bool,
-    intent: str,
-    need_aesthetics: bool = False,
-    prompt_chars: int = 0,
-) -> bool:
-    if not isinstance(rule, dict) or not rule:
-        return False
-    intent_l = str(intent or "").strip().lower()
-
-    if "empty_canvas" in rule and bool(rule.get("empty_canvas")) != bool(empty_canvas):
-        return False
-    if "has_images" in rule and bool(rule.get("has_images")) != bool(has_images):
-        return False
-    if "need_aesthetics" in rule and bool(rule.get("need_aesthetics")) != bool(
-        need_aesthetics
-    ):
-        return False
-    if "min_prompt_chars" in rule:
-        try:
-            need = int(rule.get("min_prompt_chars") or 0)
-        except (TypeError, ValueError):
-            need = 0
-        if int(prompt_chars or 0) < need:
-            return False
-
-    raw_intents = rule.get("intent_in") or rule.get("intents")
-    if raw_intents is not None:
-        if isinstance(raw_intents, str):
-            want = {p.strip().lower() for p in raw_intents.split(",") if p.strip()}
-        elif isinstance(raw_intents, list):
-            want = {str(x).strip().lower() for x in raw_intents if str(x).strip()}
-        else:
-            want = set()
-        if want and intent_l not in want:
-            return False
-        if want and not intent_l:
-            return False
-
-    keys = set(rule.keys()) & {
-        "empty_canvas",
-        "has_images",
-        "need_aesthetics",
-        "intent_in",
-        "intents",
-        "min_prompt_chars",
-    }
-    return bool(keys)
-
-
-def resolve_triggered_skill_keys(
-    *,
-    scene: str = "website",
-    empty_canvas: bool = False,
-    has_images: bool = False,
-    intent: str = "",
-    need_aesthetics: bool = False,
-    prompt_chars: int = 0,
-    already_loaded: list[str] | None = None,
-    max_n: int = 6,
-    aesthetics_triggers_only: bool = False,
-) -> list[str]:
-    loaded = {str(x).strip().lower() for x in (already_loaded or []) if str(x).strip()}
-    matched: list[dict[str, Any]] = []
-    for row in list_runtime_skills(scene=scene):
-        key = str(row.get("skillKey") or "").strip().lower()
-        if not key or key in loaded:
-            continue
-        rules = row.get("triggers") or []
-        if not rules:
-            continue
-        if aesthetics_triggers_only:
-            rules = [
-                r
-                for r in rules
-                if isinstance(r, dict) and "need_aesthetics" in r
-            ]
-            if not rules:
-                continue
-        if any(
-            _rule_matches(
-                rule,
-                empty_canvas=empty_canvas,
-                has_images=has_images,
-                intent=intent,
-                need_aesthetics=need_aesthetics,
-                prompt_chars=prompt_chars,
-            )
-            for rule in rules
-        ):
-            matched.append(row)
-    matched = _apply_mutex(matched)
-    out: list[str] = []
-    for row in matched:
-        key = str(row.get("skillKey") or "").strip().lower()
-        if not key or key in loaded:
-            continue
-        out.append(key)
-        loaded.add(key)
-        if len(out) >= max_n:
-            break
-    return out
-
-
-def _triggers_json(item: dict[str, Any]) -> str:
-    return json.dumps(_parse_triggers(item.get("triggers")), ensure_ascii=False)
-
-
-def _preferred_json(item: dict[str, Any]) -> str:
-    preferred = item.get("preferred_tools") or item.get("preferredTools") or []
-    return json.dumps(preferred if isinstance(preferred, list) else [], ensure_ascii=False)
-
-
-def _allowed_resources_json(item: dict[str, Any], *, source: str) -> str | None:
-    raw = item.get("allowed_resources")
-    if raw is None:
-        raw = item.get("allowedResources")
-    parsed = _parse_allowed_resources(raw)
-    if parsed is None:
-        if source == SOURCE_ADMIN:
-            return json.dumps(["tools"], ensure_ascii=False)
-        if source in (SOURCE_SEED, SOURCE_FILE):
-            return json.dumps(
-                ["knowledge", "aesthetics", "tools"], ensure_ascii=False
-            )
-        return None
-    return json.dumps(parsed, ensure_ascii=False)
-
-
-def _schema_json(item: dict[str, Any], *keys: str) -> str | None:
-    for k in keys:
-        if k in item and item.get(k) is not None:
-            obj, errs = validate_skill_io_schema(item.get(k), field=k)
-            if errs or obj is None:
-                return None if obj is None else json.dumps(obj, ensure_ascii=False)
-            return json.dumps(obj, ensure_ascii=False)
-    return None
-
-
-def _upsert_owned_skill(
-    session: Session,
-    item: dict[str, Any],
-    *,
-    source: str,
-    now: float,
-    skip_sources: frozenset[str],
-) -> None:
-    from app.models import DesignSkill
-
-    key = str(item.get("skill_key") or "").strip()
-    if not key:
-        return
-    meta_errs = validate_skill_meta({**item, "skill_key": key}, source=source)
-    if meta_errs:
-        logger.warning("skip skill upsert %s (%s): %s", key, source, ",".join(meta_errs))
-        return
-    namespace = _normalize_namespace(item.get("namespace"), source=source)
-    name = str(item.get("name") or key).strip() or key
-    category = str(item.get("category") or "agent").strip() or "agent"
-    when = str(item.get("when_to_use") or item.get("whenToUse") or "").strip()
-    pos = str(item.get("prompt_positive") or item.get("promptPositive") or "")
-    neg = str(item.get("prompt_negative") or item.get("promptNegative") or "")
-    scenes = str(item.get("scenes") or "all").strip() or "all"
-    sort_weight = int(item.get("sort_weight") or item.get("sortWeight") or 0)
-    mutex = str(item.get("mutex_group") or item.get("mutexGroup") or "").strip()
-    version = int(item.get("version") or 1)
-    pack_version = str(item.get("pack_version") or item.get("packVersion") or version).strip()
-    description = str(item.get("description") or "").strip()
-    logo = str(item.get("logo") or "").strip()
-    locales_obj = item.get("locales") if isinstance(item.get("locales"), dict) else {}
-    try:
-        locales_json = json.dumps(locales_obj, ensure_ascii=False) if locales_obj else ""
-    except Exception:
-        locales_json = ""
-    preferred_json = _preferred_json(item)
-    triggers_json = _triggers_json(item)
-    allowed_json = _allowed_resources_json(item, source=source)
-    input_schema_json = _schema_json(item, "input_schema", "inputSchema")
-    output_schema_json = _schema_json(item, "output_schema", "outputSchema")
-    owner_user_id = str(item.get("owner_user_id") or item.get("ownerUserId") or "").strip() or None
-
-    row = crud.get_design_skill_by_key(session=session, skill_key=key)
-    skill_id: int | None = None
-    next_ver = version
-    if row:
-        src = _normalize_source(_row_get(row, "source"), default=source)
-        # Community seed is cold-start only: never update / never reclaim admin rows.
-        if source == SOURCE_SEED:
-            return
-        if src in skip_sources:
-            return
-        try:
-            cur_ver = int(_row_get(row, "version") or 0)
-        except (TypeError, ValueError):
-            cur_ver = 0
-        next_ver = (
-            max(version, cur_ver + 1)
-            if source in (SOURCE_SEED, SOURCE_FILE)
-            else max(version, 1)
-        )
-        skill_id = int(row.id or 0) or None
-        row.name = name
-        row.category = category
-        row.prompt_positive = pos
-        row.prompt_negative = neg
-        row.when_to_use = when
-        row.preferred_tools = preferred_json
-        if allowed_json is not None:
-            row.allowed_resources = allowed_json
-        row.triggers = triggers_json
-        row.mutex_group = mutex
-        row.version = next_ver
-        row.pack_version = pack_version
-        row.description = description
-        row.logo = logo
-        row.locales = locales_json
-        row.source = source
-        row.namespace = namespace
-        if owner_user_id:
-            row.owner_user_id = owner_user_id
-        if input_schema_json is not None:
-            row.input_schema = input_schema_json
-        if output_schema_json is not None:
-            row.output_schema = output_schema_json
-        row.sort_weight = sort_weight
-        row.scenes = scenes
-        row.enabled = 1
-        row.updated_at = now
-        session.add(row)
-    else:
-        default_allowed = allowed_json or json.dumps(
-            ["knowledge", "aesthetics", "tools"]
-            if source != SOURCE_ADMIN
-            else ["tools"],
-            ensure_ascii=False,
-        )
-        row = DesignSkill(
-            skill_key=key,
-            name=name,
-            category=category,
-            prompt_positive=pos,
-            prompt_negative=neg,
-            when_to_use=when,
-            preferred_tools=preferred_json,
-            allowed_resources=default_allowed,
-            triggers=triggers_json,
-            mutex_group=mutex,
-            version=version,
-            pack_version=pack_version,
-            description=description,
-            logo=logo,
-            locales=locales_json,
-            source=source,
-            namespace=namespace,
-            owner_user_id=owner_user_id,
-            input_schema=input_schema_json,
-            output_schema=output_schema_json,
-            sort_weight=sort_weight,
-            scenes=scenes,
-            default_model="doubao",
-            max_retries=2,
-            enabled=1,
-            output_format="json",
-            allow_user_model_override=0,
-            created_at=now,
-            updated_at=now,
-        )
-        session.add(row)
-        session.flush()
-        skill_id = int(row.id or 0) or None
-        next_ver = version
-
-    if skill_id:
-        save_skill_revision(
-            session,
-            skill_id=skill_id,
-            item={
-                "skillKey": key,
-                "name": name,
-                "description": description,
-                "category": category,
-                "whenToUse": when,
-                "promptPositive": pos,
-                "promptNegative": neg,
-                "preferredTools": json.loads(preferred_json) if preferred_json else [],
-                "allowedResources": json.loads(allowed_json) if allowed_json else None,
-                "inputSchema": json.loads(input_schema_json) if input_schema_json else None,
-                "outputSchema": json.loads(output_schema_json) if output_schema_json else None,
-                "triggers": json.loads(triggers_json) if triggers_json else [],
-                "mutexGroup": mutex or None,
-                "version": next_ver,
-                "packVersion": pack_version,
-                "namespace": namespace,
-                "source": source,
-                "scenes": scenes,
-                "sortWeight": sort_weight,
-            },
-        )
-
-
-def _skills_disk_signature() -> str:
-    parts: list[str] = []
-    try:
-        seed_path = _skills_seed_path()
-        if seed_path.is_file():
-            st = seed_path.stat()
-            parts.append(f"seed:{st.st_mtime_ns}:{st.st_size}")
-    except Exception:
-        parts.append("seed:missing")
-    for root in _file_skills_dirs():
-        for pack in sorted(p for p in root.iterdir() if p.is_dir()):
-            try:
-                meta_m = 0
-                body_m = 0
-                for name in _META_NAMES:
-                    mp = pack / name
-                    if mp.is_file():
-                        meta_m = mp.stat().st_mtime_ns
-                        break
-                sp = _skill_md_path(pack)
-                if sp is not None:
-                    body_m = sp.stat().st_mtime_ns
-                parts.append(f"{root.name}/{pack.name}:{meta_m}:{body_m}")
-            except Exception:
-                parts.append(f"{root.name}/{pack.name}:err")
-    return "|".join(parts)
-
-
-def ensure_design_skills(*, force: bool = False) -> None:
-    """Upsert seed + file skills; never overwrite admin."""
-    global _SKILLS_READY, _DISK_SIGNATURE
-    if _SKILLS_READY and not force:
-        return
-    with _SKILLS_LOCK:
-        if _SKILLS_READY and not force:
-            return
-        now = time.time()
-        from app.services.db import init_schema
-        from app.services.design.admin.schema import ensure_design_tables_boot
-
-        init_schema()
-        ensure_design_tables_boot()
-        with Session(core_db.engine) as session:
-            for item in _SEED:
-                seeded = dict(item)
-                seeded.setdefault("namespace", NS_CORE)
-                seeded.setdefault(
-                    "allowed_resources",
-                    ["knowledge", "aesthetics", "tools"],
-                )
-                _upsert_owned_skill(
-                    session,
-                    seeded,
-                    source=SOURCE_SEED,
-                    now=now,
-                    skip_sources=_PROTECTED_FROM_SEED,
-                )
-            for item in _load_file_skills():
-                _upsert_owned_skill(
-                    session,
-                    item,
-                    source=SOURCE_FILE,
-                    now=now,
-                    skip_sources=_PROTECTED_FROM_FILE,
-                )
-            _bump_unchanged_seed_skill_bodies(session, now=now)
-            # Retired need_* packs (design_spec/vision/aesthetics) → Skills; delete leftovers.
-            try:
-                crud.delete_design_prompt_packs_by_kinds(
-                    session=session, kinds=list(RETIRED_NEED_PROMPT_KINDS)
-                )
-            except Exception:
-                pass
-            session.commit()
-        invalidate_skill_key_cache()
-        _DISK_SIGNATURE = _skills_disk_signature()
-        _SKILLS_READY = True
-
-
-# Prior seed skill bodies that may be replaced once (Admin-owned rows are kept).
-_SEED_SKILL_BODY_PREV: dict[str, tuple[str, ...]] = {
-    "design_methodology": (
-        "# Design methodology\n\n"
-        "Step skeleton for create-from-scratch. Which op, where to place, vector vs image-gen "
-        "vs transparent bg → need_skills=[\"user.layout_ops\"]. Numeric details (palette, "
-        "banner px, …) → need_knowledge (pick from the Knowledge catalog).\n\n"
-        "## Deliverables (one line each)\n"
-        "- **website / landing**: nav → hero → benefits → CTA → footer\n"
-        "- **mobile / H5**: single column; primary CTA in the thumb zone\n"
-        "- **ecommerce hero**: ~1:1; subject largest; no nav, no forms\n"
-        "- **ecommerce detail**: hero → title/price → benefits → specs → rich text\n"
-        "- **Banner**: copy in the middle safe band\n"
-        "- **poster / roll-up**: hero visual + info group; tall top/mid/bottom\n"
-        "- **image**: subject graphic; no page chrome\n\n"
-        "## Steps\n"
-        "1. Confirm deliverable, size, required copy; ask if missing; never invent price/phone\n"
-        "2. Fixed size → create_frame first, then place layers inside (see user.layout_ops)\n"
-        "3. User attached images → need_skills includes vision_extract; product workflows "
-        "add user.* from the catalog\n"
-        "4. Node ids must come from SCENE\n",
-        "# Design methodology\n\n"
-        "Process (think → act): brief → structure → place → self-check.\n"
-        "Which op, where to place, vector vs image-gen vs transparent bg → "
-        "need_skills=[\"user.layout_ops\"]. Numeric details (palette, banner px, …) → "
-        "need_knowledge (pick from the Knowledge catalog).\n\n"
-        "## Brief (one line in thought)\n"
-        "Name deliverable, audience/job, and size. Ask if missing; never invent price/phone.\n\n"
-        "## Deliverables (one line each)\n"
-        "- **website / landing**: nav → hero → benefits → CTA → footer\n"
-        "- **mobile / H5**: single column; primary CTA in the thumb zone\n"
-        "- **ecommerce hero**: ~1:1; subject largest; no nav, no forms\n"
-        "- **ecommerce detail**: hero → title/price → benefits → specs → rich text\n"
-        "- **Banner**: copy in the middle safe band\n"
-        "- **poster / roll-up**: hero visual + info group; tall top/mid/bottom\n"
-        "- **image**: subject graphic; no page chrome\n\n"
-        "## Steps\n"
-        "1. Confirm deliverable, size, required copy; ask if missing; never invent price/phone\n"
-        "2. Fixed size → create_frame first, then place layers inside (see user.layout_ops)\n"
-        "3. Hierarchy: one signature focus; keep surroundings quiet (avoid generic AI template looks)\n"
-        "4. User attached images → need_skills includes vision_extract; product workflows "
-        "add user.* from the catalog\n"
-        "5. Node ids must come from SCENE\n"
-        "6. Before done: quick self-check — hierarchy, safe margins, copy language match\n",
-        "# 设计方法论\n\n"
-        "从零创作的步骤骨架。元素用什么 op、放哪、矢量/生图/透明底 → need_skills=[\"user.layout_ops\"]。"
-        "配色/Banner 像素等数字细则 → need_knowledge（从 Knowledge 目录选）。\n\n"
-        "## 物料（一句话）\n"
-        "- **website / 落地页**：导航 → 英雄 → 卖点 → CTA → 页脚\n"
-        "- **mobile / H5**：单列；主 CTA 在易触区\n"
-        "- **电商主图**：约 1:1；主体最大；无导航无表单\n"
-        "- **电商详情**：主图 → 标题价格 → 卖点 → 规格 → 图文\n"
-        "- **Banner**：文案在中部安全带\n"
-        "- **poster / 易拉宝**：主视觉 + 信息组；竖长上中下\n"
-        "- **image**：主体图形；无页面壳\n\n"
-        "## 步骤\n"
-        "1. 确认物料、尺寸、必出文案；缺则 ask；未提供的价格/电话不写\n"
-        "2. 固定尺寸先 create_frame，再板内落层（细则见 user.layout_ops）\n"
-        "3. 有附图 → need_skills 含 vision_extract；产品工作流（需求整理等）按目录再加 user.*\n"
-        "4. 节点 id 必须来自 SCENE\n",
-        "# Design methodology\n\n"
-        "Process (think → act): brief → structure → place → self-check.\n"
-        "This skill is the **step skeleton only** — not typography/aesthetic encyclopedia.\n"
-        "Visual craft (type mood, contrast, overflow, decoration): `need_skills` entries from the "
-        "**Skills catalog** whose when_to_use matches (enabled skills only; never invent skill keys).\n"
-        "Layout/op cheat-sheets: same — pick from catalog if present.\n"
-        "Numeric details (palette, banner px, …) → need_knowledge (Knowledge catalog).\n\n"
-        "## Brief (one line in thought)\n"
-        "Name deliverable, audience/job, and size. Ask if missing; never invent price/phone.\n\n"
-        "## Deliverables (one line each)\n"
-        "- **website / landing**: nav → hero → benefits → CTA → footer\n"
-        "- **mobile / H5**: single column; primary CTA in the thumb zone\n"
-        "- **ecommerce hero**: ~1:1; subject largest; no nav, no forms\n"
-        "- **ecommerce detail**: hero → title/price → benefits → specs → rich text\n"
-        "- **Banner**: copy in the middle safe band\n"
-        "- **poster / roll-up**: hero visual + info group; tall top/mid/bottom\n"
-        "- **image**: subject graphic; no page chrome\n\n"
-        "## Steps\n"
-        "1. Confirm deliverable, size, required copy; ask if missing; never invent price/phone\n"
-        "2. Fixed size → create_frame first, then place layers inside\n"
-        "3. Before heavy place: `need_skills` any enabled catalog skills for type/layout/aesthetics "
-        "that match the job\n"
-        "4. Hierarchy: one signature focus; keep surroundings quiet (avoid generic AI template looks)\n"
-        "5. User attached images → need_skills includes vision_extract; product workflows "
-        "add matching user.* from the catalog\n"
-        "6. Type: if style exists in Available fonts → create_text + fontFamily; if not "
-        "(brush/3D/effect lettering) or font mood mismatches the art → create_image + letteringText\n"
-        "7. Node ids must come from SCENE\n"
-        "8. Before done: self-check — no clipped titles, contrast OK, no junk SVG/emoji tofu, "
-        "copy language match\n",
-        "# Design methodology\n\n"
-        "Process (think → act): brief → structure → place → self-check.\n"
-        "This skill is the **step skeleton only** — not typography/aesthetic encyclopedia.\n"
-        "Visual craft (type mood, contrast, overflow, decoration): `need_skills` entries from the "
-        "**Skills catalog** whose when_to_use matches (enabled skills only; never invent skill keys).\n"
-        "Layout/op cheat-sheets: same — pick from catalog if present.\n"
-        "Numeric details (palette, banner px, …) → need_knowledge (Knowledge catalog).\n\n"
-        "## Brief (one line in thought)\n"
-        "Name deliverable, audience/job, and size. Ask if missing; never invent price/phone.\n\n"
-        "## Deliverables (one line each)\n"
-        "- **website / landing**: nav → hero → benefits → CTA → footer\n"
-        "- **mobile / H5**: single column; primary CTA in the thumb zone\n"
-        "- **ecommerce hero**: ~1:1; subject largest; no nav, no forms\n"
-        "- **ecommerce detail**: hero → title/price → benefits → specs → rich text\n"
-        "- **Banner**: copy in the middle safe band\n"
-        "- **poster / roll-up**: **create_image genPrompt hero** + info group; tall top/mid/bottom — "
-        "do not fake the hero with shape piles\n"
-        "- **image**: subject graphic; no page chrome\n\n"
-        "## Steps\n"
-        "1. Confirm deliverable, size, required copy; ask if missing; never invent price/phone\n"
-        "2. Fixed size → create_frame first, then place layers inside\n"
-        "3. Before heavy place: `need_skills` any enabled catalog skills for type/layout/aesthetics "
-        "that match the job\n"
-        "4. Poster / festive / illustrated jobs: first full-bleed (or near) `create_image` hero, "
-        "then text layers\n"
-        "5. Hierarchy: one signature focus; keep surroundings quiet (avoid generic AI template looks)\n"
-        "6. User attached images → need_skills includes vision_extract; product workflows "
-        "add matching user.* from the catalog\n"
-        "7. Type: if style exists in Available fonts → create_text + fontFamily; if not "
-        "(brush/3D/effect lettering) or font mood mismatches the art → create_image + letteringText\n"
-        "8. Node ids must come from SCENE\n"
-        "9. Before done: self-check — hero image present for posters, no clipped titles, contrast OK, "
-        "no junk SVG/emoji tofu, copy language match\n",
-    ),
-    "canvas_edit": (
-        "# Canvas edit\n\n"
-        "Add shape / recolor / rewrite via tool_ops; use sensible defaults for missing size/color.\n"
-        "Change type with update_node; keep id.\n"
-        "Node ids only from SCENE.\n"
-        "Hero: keep subject and whitespace; PDP: keep module order; Banner: keep copy in the safe band.\n",
-    ),
-    "vision_extract": (
-        "# Vision extract\n\n"
-        "Only describe what is visible. Label deliverable: hero / PDP / Banner / poster / "
-        "roll-up / web or App / illustration icon.\n\n"
-        "List: regions, key elements and rough placement, 3–5 main colors, suggested size or aspect.\n"
-        "Hero: subject + badges; PDP: module sections; Banner: whether copy sits mid-band; "
-        "tall art: top/mid/bottom.\n",
-    ),
-    "image_gen": (
-        "# Image generation\n\n"
-        "`create_image` `genPrompt` (or `prompt`): natural-language brief for the image model "
-        "to make a bitmap — not an editable canvas text node.\n\n"
-        "## Choose (your judgment; no fixed copy→op checklist)\n"
-        "- **create_image**: photo / illustration / material, or text that is a visual effect "
-        "(calligraphy strokes, color-block type, extrusion — fonts cannot do this) → generate "
-        "as one graphic; prompt must state the glyphs, style, palette, transparent bg or not.\n"
-        "- **create_text**: editable, rewritable layout text (not the effects above).\n\n"
-        "If the user forbids image gen, skip create_image; fall back to plain create_text for "
-        "effect lettering — do not pretend a calligraphy font exists.\n"
-        "Stylized lettering may set letteringText (exact on-image string) so image_process "
-        "kind=replaceText (meta.originalText+newText) can swap words.\n"
-        "Do not invent facts into the image.\n"
-        "Size, transparent bg, stacking → user.layout_ops (do not expand here).\n",
-        "# Image generation\n\n"
-        "`create_image` `genPrompt` (or `prompt`): natural-language brief for the image model "
-        "to make a bitmap — not an editable canvas text node.\n\n"
-        "## Choose (your judgment; no fixed copy→op checklist)\n"
-        "- **create_image**: photo / illustration / material, or text that is a visual effect "
-        "(calligraphy strokes, color-block type, extrusion — fonts cannot do this) → generate "
-        "as one graphic; prompt must state the glyphs, style, palette, transparent bg or not.\n"
-        "- **create_text**: editable, rewritable layout text (not the effects above).\n\n"
-        "If the user forbids image gen, skip create_image; fall back to plain create_text for "
-        "effect lettering — do not pretend a calligraphy font exists.\n"
-        "Stylized lettering: set letteringText to the exact on-image string AND put those same "
-        "glyphs in genPrompt (hydrate forwards letteringText into the image model).\n"
-        "Do not invent facts into the image.\n"
-        "Size, transparent bg, stacking → user.layout_ops (do not expand here).\n",
-        "# Image generation\n\n"
-        "`create_image` `genPrompt` (or `prompt`): natural-language brief for the image model "
-        "to make a bitmap — not an editable canvas text node.\n\n"
-        "## Choose (your judgment; no fixed copy→op checklist)\n"
-        "- **create_text + fontFamily**: UI / body / title when the style exists in Available fonts.\n"
-        "- **create_image**: photo / illustration / material, OR text that fonts cannot do "
-        "(brush calligraphy, color-block type, extrusion) → generate as one graphic; set "
-        "letteringText + put those glyphs in genPrompt; ask for isolated glyphs on a plain "
-        "solid plate (hydrate auto-cutouts to transparent PNG).\n\n"
-        "If the user forbids image gen, skip create_image; fall back to plain create_text for "
-        "effect lettering — do not pretend a calligraphy font exists.\n"
-        "Do not invent facts into the image.\n"
-        "Size, transparent bg, stacking → need_skills from Skills catalog for layout/ops "
-        "(do not expand here).\n",
-    ),
-}
-
-
-def _norm_skill_body(text: str) -> str:
-    return str(text or "").replace("\r\n", "\n").strip()
-
-
-def _bump_unchanged_seed_skill_bodies(session: Any, *, now: float) -> None:
-    """Replace listed seed skills when body matches a prior seed, or seed version is newer."""
-    seed_by_key = {
-        str(it.get("skill_key") or "").strip(): it
-        for it in _SEED
-        if isinstance(it, dict) and str(it.get("skill_key") or "").strip()
-    }
-    for key in _SEED_SKILL_BODY_PREV:
-        seed_item = seed_by_key.get(key) or {}
-        new_pos = str(seed_item.get("prompt_positive") or "").strip()
-        new_neg = str(seed_item.get("prompt_negative") or "")
-        if not new_pos:
-            continue
-        row = crud.get_design_skill_by_key(session=session, skill_key=key)
-        if not row:
-            continue
-        src = _normalize_source(_row_get(row, "source"), default=SOURCE_SEED)
-        if src != SOURCE_SEED:
-            continue
-        try:
-            seed_ver = int(seed_item.get("version") or 0)
-            cur_ver = int(_row_get(row, "version") or 0)
-        except (TypeError, ValueError):
-            seed_ver, cur_ver = 0, 0
-        cur = _norm_skill_body(str(_row_get(row, "prompt_positive") or ""))
-        new_norm = _norm_skill_body(new_pos)
-        prev_norms = {
-            _norm_skill_body(p) for p in (_SEED_SKILL_BODY_PREV.get(key) or ())
-        }
-        version_bump = seed_ver > cur_ver
-        body_match = cur in prev_norms
-        if cur == new_norm:
-            continue
-        if not version_bump and not body_match:
-            continue
-        row.prompt_positive = new_pos
-        if new_neg:
-            row.prompt_negative = new_neg
-        row.version = max(cur_ver, seed_ver)
-        row.updated_at = now
-        session.add(row)
-
-
-def stop_skills_hot_reload() -> None:
-    global _HOT_RELOAD_THREAD
-    _HOT_RELOAD_STOP.set()
-    t = _HOT_RELOAD_THREAD
-    _HOT_RELOAD_THREAD = None
-    if t and t.is_alive() and t is not threading.current_thread():
-        try:
-            t.join(timeout=1.5)
-        except Exception:
-            pass
-
-
-def start_skills_hot_reload() -> bool:
-    """Poll seed + file-pack mtimes and force-resync when they change."""
-    global _HOT_RELOAD_THREAD
-    try:
-        from app.core.config import settings
-
-        enabled = bool(getattr(settings, "design_skills_hot_reload", True))
-        interval = float(getattr(settings, "design_skills_hot_reload_interval_sec", 2.0) or 2.0)
-    except Exception:
-        enabled = True
-        interval = 2.0
-    if not enabled:
-        return False
-    interval = max(0.5, min(interval, 60.0))
-    stop_skills_hot_reload()
-    _HOT_RELOAD_STOP.clear()
-
-    def _loop() -> None:
-        global _DISK_SIGNATURE
-        while not _HOT_RELOAD_STOP.wait(interval):
-            try:
-                sig = _skills_disk_signature()
-                prev = _DISK_SIGNATURE
-                if prev is None:
-                    _DISK_SIGNATURE = sig
-                    continue
-                if sig != prev:
-                    logger.info("design skills disk change detected — hot reload")
-                    ensure_design_skills(force=True)
-            except Exception:
-                logger.exception("design skills hot reload failed")
-
-    t = threading.Thread(target=_loop, name="design-skills-hot-reload", daemon=True)
-    _HOT_RELOAD_THREAD = t
-    t.start()
-    return True
-
-
-def reload_skills_if_disk_changed() -> bool:
-    """One-shot check used by tests / admin; returns True if reloaded."""
-    global _DISK_SIGNATURE
-    sig = _skills_disk_signature()
-    if _DISK_SIGNATURE is not None and sig == _DISK_SIGNATURE:
-        return False
-    ensure_design_skills(force=True)
-    return True
-
-
-_SLUG_RE = re.compile(r"[^a-z0-9]+")
-
-
-def _slug_local_key(name: str) -> str:
-    s = _SLUG_RE.sub("-", str(name or "").strip().lower()).strip("-")
-    return (s or "skill")[:48]
-
-
-def _picker_card(row: dict[str, Any], *, user_id: str) -> dict[str, Any]:
-    uid = str(user_id or "").strip()
-    owner = str(row.get("ownerUserId") or "").strip() or None
-    return {
-        "id": int(row.get("id") or 0),
-        "skillKey": row.get("skillKey"),
-        "qualifiedKey": row.get("qualifiedKey"),
-        "name": str(row.get("name") or ""),
-        "description": str(row.get("description") or ""),
-        "whenToUse": str(row.get("whenToUse") or ""),
-        "logo": row.get("logo"),
-        "namespace": row.get("namespace"),
-        "source": row.get("source"),
-        "ownerUserId": owner,
-        "category": str(row.get("category") or ""),
-        "mine": bool(owner and uid and owner == uid),
-        "triggers": list(row.get("triggers") or []),
-        "enabled": bool(row.get("enabled")),
-    }
-
-
-def list_skills_for_picker(
-    *, user_id: str, scene: str = "website"
-) -> list[dict[str, Any]]:
-    """Slim catalog for `/` picker + Skills tab (bodies omitted)."""
-    uid = str(user_id or "").strip()
-    if not uid:
-        return []
-    rows = list_runtime_skills(scene=scene, user_id=uid, enabled_only=True)
-    return [_picker_card(r, user_id=uid) for r in rows]
-
-
-def list_skills_for_manage(
-    *, user_id: str, scene: str = "website"
-) -> list[dict[str, Any]]:
-    """Toolbox list: mine (all) + official; includes user-disabled with enabled=false."""
-    uid = str(user_id or "").strip()
-    if not uid:
-        return []
-    prefs = _load_user_skill_prefs(uid)
-    # Globally enabled rows + this user's mine (even if row.enabled=0).
-    runtime_on = list_runtime_skills(scene=scene, user_id=uid, enabled_only=False)
-    out: list[dict[str, Any]] = []
-    seen: set[int] = set()
-    for item in runtime_on:
-        sid = int(item.get("id") or 0)
-        key = str(item.get("skillKey") or "").strip()
-        owner = str(item.get("ownerUserId") or "").strip() or None
-        mine = bool(owner and owner == uid)
-        globally_on = bool(item.get("enabled"))
-        if not mine and not globally_on:
-            continue
-        if sid and sid in seen:
-            continue
-        if sid:
-            seen.add(sid)
-        card = _picker_card(item, user_id=uid)
-        # Preview body for toolbox (mine + official).
-        card["promptPositive"] = str(item.get("promptPositive") or "")
-        card["promptNegative"] = str(item.get("promptNegative") or "")
-        if mine:
-            # Own row.enabled is authoritative; pref can still force off.
-            card["enabled"] = globally_on and prefs.get(key.lower(), True)
-        else:
-            card["enabled"] = prefs.get(key.lower(), True)
-        out.append(card)
-    mine_cards = [c for c in out if c.get("mine")]
-    official_cards = [c for c in out if not c.get("mine")]
-    return mine_cards + official_cards
-
-
-def list_my_skills(*, user_id: str) -> list[dict[str, Any]]:
-    """User-owned skills for management (includes disabled; includes body)."""
-    uid = str(user_id or "").strip()
-    if not uid:
-        return []
-    ensure_design_skills()
-    prefs = _load_user_skill_prefs(uid)
-    with Session(core_db.engine) as session:
-        rows = crud.list_design_skills_by_owner(
-            session=session, owner_user_id=uid, namespace=NS_USER
-        )
-    out: list[dict[str, Any]] = []
-    for r in rows:
-        item = _pub(r)
-        card = _picker_card(item, user_id=uid)
-        key = str(item.get("skillKey") or "").strip().lower()
-        card["enabled"] = bool(item.get("enabled")) and prefs.get(key, True)
-        card["promptPositive"] = str(item.get("promptPositive") or "")
-        card["promptNegative"] = str(item.get("promptNegative") or "")
-        out.append(card)
-    return out
-
-
-def set_user_skill_enabled(
-    *, user_id: str, skill_id: int, enabled: bool
-) -> dict[str, Any]:
-    """Toggle a skill for this user. Own skills update row; others use prefs."""
-    uid = str(user_id or "").strip()
-    if not uid:
-        raise ValueError("user_id required")
-    sid = int(skill_id or 0)
-    if sid <= 0:
-        raise ValueError("skill_id required")
-    ensure_design_skills()
-    with Session(core_db.engine) as session:
-        row = crud.get_design_skill(session=session, item_id=sid)
-        if not row:
-            raise ValueError("skill not found")
-        item = _pub(row)
-        key = str(item.get("skillKey") or "").strip()
-        if not key:
-            raise ValueError("skill_key missing")
-        owner = str(item.get("ownerUserId") or "").strip() or None
-        mine = bool(owner and owner == uid)
-        ns = str(item.get("namespace") or "")
-        # Non-owners can only toggle non-user-namespace (official/core/ext) skills.
-        if not mine:
-            if ns == NS_USER and owner:
-                raise ValueError("not_skill_owner")
-            if not bool(item.get("enabled")):
-                raise ValueError("skill_disabled_globally")
-        if mine:
-            crud.update_design_skill_enabled(
-                session=session,
-                item_id=sid,
-                owner_user_id=uid,
-                enabled=1 if enabled else 0,
-                commit=False,
-            )
-        crud.upsert_user_skill_pref(
-            session=session,
-            user_id=uid,
-            skill_key=key,
-            enabled=1 if enabled else 0,
-            commit=False,
-        )
-        session.commit()
-        saved = crud.get_design_skill(session=session, item_id=sid)
-    invalidate_skill_key_cache()
-    if not saved:
-        raise RuntimeError("toggle skill failed")
-    card = _picker_card(_pub(saved), user_id=uid)
-    card["enabled"] = bool(enabled)
-    if mine:
-        pub = _pub(saved)
-        card["promptPositive"] = str(pub.get("promptPositive") or "")
-        card["promptNegative"] = str(pub.get("promptNegative") or "")
-    return card
-
-
-def resolve_accessible_skill_keys(
-    *,
-    user_id: str,
-    refs: list[Any] | None,
-    scene: str = "website",
-    max_n: int = 8,
-) -> list[str]:
-    """Map chip refs (skillKey / id / qualified) → storage keys visible to user."""
-    uid = str(user_id or "").strip()
-    if not uid or not refs:
-        return []
-    runtime = list_runtime_skills(scene=scene, user_id=uid, enabled_only=True)
-    by_key = {
-        str(r.get("skillKey") or "").strip().lower(): str(r.get("skillKey") or "").strip()
-        for r in runtime
-        if str(r.get("skillKey") or "").strip()
-    }
-    by_id = {
-        int(r.get("id") or 0): str(r.get("skillKey") or "").strip()
-        for r in runtime
-        if int(r.get("id") or 0) > 0 and str(r.get("skillKey") or "").strip()
-    }
-    out: list[str] = []
-    seen: set[str] = set()
-    for raw in refs:
-        if len(out) >= max_n:
-            break
-        s = str(raw or "").strip()
-        if not s:
-            continue
-        if s.lower().startswith("skill:"):
-            s = s[6:].strip()
-        storage = ""
-        if s.isdigit():
-            storage = by_id.get(int(s)) or ""
-        if not storage:
-            hit = resolve_storage_skill_key(s, scene=scene)
-            if hit and hit.lower() in by_key:
-                storage = by_key[hit.lower()]
-        if not storage:
-            low = s.lower()
-            if low in by_key:
-                storage = by_key[low]
-        if not storage:
-            continue
-        low = storage.lower()
-        if low in seen:
-            continue
-        seen.add(low)
-        out.append(storage)
-    return out
-
-
-def upsert_end_user_skill(*, user_id: str, payload: dict[str, Any]) -> dict[str, Any]:
-    """Create/update a user-owned skill (namespace=user, source=admin)."""
-    uid = str(user_id or "").strip()
-    if not uid:
-        raise ValueError("user_id required")
-    name = str(payload.get("name") or "").strip()
-    if not name:
-        raise ValueError("name required")
-    prompt_positive = str(
-        payload.get("promptPositive") or payload.get("prompt_positive") or ""
-    ).strip()
-    if not prompt_positive:
-        raise ValueError("promptPositive required")
-    prompt_negative = str(
-        payload.get("promptNegative") or payload.get("prompt_negative") or ""
-    ).strip()
-    when_to_use = str(
-        payload.get("whenToUse") or payload.get("when_to_use") or ""
-    ).strip()
-    description = str(payload.get("description") or "").strip()
-    logo = str(payload.get("logo") or "").strip() or None
-    category = str(payload.get("category") or "custom").strip() or "custom"
-    enabled = 1 if payload.get("enabled", True) else 0
-    sid_raw = payload.get("id")
-    sid = int(sid_raw) if sid_raw not in (None, "", 0, "0") else None
-
-    triggers_raw = payload.get("triggers")
-    if isinstance(triggers_raw, list):
-        triggers_json = json.dumps(triggers_raw, ensure_ascii=False)
-    elif isinstance(triggers_raw, str) and triggers_raw.strip():
-        triggers_json = triggers_raw.strip()
-    else:
-        # Auto trigger from name tokens for `/` search.
-        triggers_json = json.dumps(
-            [{"type": "keyword", "value": name}],
-            ensure_ascii=False,
-        )
-
-    ensure_design_skills()
-    with Session(core_db.engine) as session:
-        if sid:
-            row = crud.get_design_skill(session=session, item_id=sid)
-            if not row:
-                raise ValueError("skill not found")
-            owner = str(_row_get(row, "owner_user_id") or "").strip()
-            if owner != uid:
-                raise ValueError("not_skill_owner")
-            skill_key = str(_row_get(row, "skill_key") or "").strip()
-            try:
-                cur_ver = int(_row_get(row, "version") or 1)
-            except Exception:
-                cur_ver = 1
-            next_ver = cur_ver + 1
-            meta_errs = validate_skill_meta(
-                {
-                    "skill_key": skill_key,
-                    "name": name,
-                    "prompt_positive": prompt_positive,
-                    "allowed_resources": ["tools"],
-                    "namespace": NS_USER,
-                },
-                source=SOURCE_ADMIN,
-            )
-            if meta_errs:
-                raise ValueError("; ".join(meta_errs))
-            saved = crud.update_end_user_design_skill(
-                session=session,
-                item_id=sid,
-                owner_user_id=uid,
-                name=name,
-                category=category,
-                prompt_positive=prompt_positive,
-                prompt_negative=prompt_negative or None,
-                when_to_use=when_to_use or None,
-                description=description or None,
-                logo=logo,
-                triggers_json=triggers_json,
-                version=next_ver,
-                enabled=enabled,
-            )
-        else:
-            raw_key = str(
-                payload.get("skillKey") or payload.get("skill_key") or ""
-            ).strip()
-            if raw_key:
-                _, local = split_namespace_key(raw_key)
-                local = local or raw_key
-            else:
-                local = f"{_slug_local_key(name)}-{uid[:8]}"
-            skill_key = qualify_skill_key(NS_USER, local)
-            meta_errs = validate_skill_meta(
-                {
-                    "skill_key": skill_key,
-                    "name": name,
-                    "prompt_positive": prompt_positive,
-                    "allowed_resources": ["tools"],
-                    "namespace": NS_USER,
-                },
-                source=SOURCE_ADMIN,
-            )
-            if meta_errs:
-                raise ValueError("; ".join(meta_errs))
-            exists = crud.get_design_skill_by_key(
-                session=session, skill_key=skill_key
-            )
-            if exists:
-                raise ValueError(f"skill_key_taken:{skill_key}")
-            saved = crud.insert_end_user_design_skill(
-                session=session,
-                skill_key=skill_key,
-                name=name,
-                category=category,
-                prompt_positive=prompt_positive,
-                prompt_negative=prompt_negative or None,
-                when_to_use=when_to_use or None,
-                triggers_json=triggers_json,
-                source=SOURCE_ADMIN,
-                namespace=NS_USER,
-                owner_user_id=uid,
-                description=description or None,
-                logo=logo,
-                enabled=enabled,
-            )
-    invalidate_skill_key_cache()
-    if not saved:
-        raise RuntimeError("upsert skill failed")
-    item = _pub(saved)
-    card = _picker_card(item, user_id=uid)
-    card["promptPositive"] = str(item.get("promptPositive") or "")
-    card["promptNegative"] = str(item.get("promptNegative") or "")
-    return card
-
-
-def delete_end_user_skill(*, user_id: str, skill_id: int) -> bool:
-    uid = str(user_id or "").strip()
-    if not uid:
-        return False
-    with Session(core_db.engine) as session:
-        row = crud.get_design_skill(session=session, item_id=int(skill_id))
-        if not row:
-            return False
-        owner = str(_row_get(row, "owner_user_id") or "").strip()
-        if owner != uid:
-            raise ValueError("not_skill_owner")
-        src = _normalize_source(_row_get(row, "source"), default=SOURCE_ADMIN)
-        if src in (SOURCE_SEED, SOURCE_FILE):
-            raise ValueError("cannot_delete_system_skill")
-        ok = crud.delete_owned_design_skill(
-            session=session, item_id=int(skill_id), owner_user_id=uid
-        )
-    if ok:
-        invalidate_skill_key_cache()
-    return ok
-
-
-_MAX_USER_SKILL_ZIP_BYTES = 5 * 1024 * 1024
-_MAX_USER_SKILL_ZIP_UNCOMPRESSED = 20 * 1024 * 1024
-_ZIP_LOGO_EXTS = (".png", ".svg", ".webp", ".jpg", ".jpeg")
-
-
-def _zip_check(cid: str, ok: bool, label: str, detail: str | None = None) -> dict[str, Any]:
-    out: dict[str, Any] = {"id": cid, "ok": ok, "label": label}
-    if detail:
-        out["detail"] = detail
-    return out
-
-
-def _import_result(
-    *,
-    status: str,
-    file_name: str,
-    checks: list[dict[str, Any]],
-    errors: list[str] | None = None,
-    item: dict[str, Any] | None = None,
-    existing: dict[str, Any] | None = None,
-    scan_ok: bool | None = None,
-) -> dict[str, Any]:
-    ok = bool(scan_ok) if scan_ok is not None else status in ("ok", "exists")
-    return {
-        "status": status,
-        "fileName": file_name,
-        "scan": {"ok": ok, "checks": checks, "errors": list(errors or [])},
-        "item": item,
-        "existing": existing,
-    }
-
-
-def _zip_entry_allowed(rel: str) -> bool:
-    """Whitelist pack files inside a user-uploaded skill zip."""
-    name = Path(rel.replace("\\", "/")).name.lower()
-    if name in _META_NAMES or name in ("skill.md", "design.md"):
-        return True
-    if name in ("license", "license.txt", "license.md", "licence", "licence.txt"):
-        return True
-    if name.startswith(".") or name.startswith("__macosx"):
-        return False
-    return name.endswith(_ZIP_LOGO_EXTS)
-
-
-def _safe_zip_relpath(name: str) -> str | None:
-    raw = str(name or "").replace("\\", "/").strip()
-    if not raw or raw.endswith("/"):
-        return None
-    if raw.startswith("/") or re.match(r"^[a-zA-Z]:", raw):
-        return None
-    parts = [p for p in raw.split("/") if p and p != "."]
-    if not parts or any(p == ".." for p in parts):
-        return None
-    if parts[0].lower() == "__macosx":
-        return None
-    return "/".join(parts)
-
-
-def _resolve_extracted_pack_dir(dest: Path, written: list[str]) -> Path:
-    tops = {p.split("/", 1)[0] for p in written}
-    if len(tops) != 1:
-        return dest
-    candidate = dest / next(iter(tops))
-    return candidate if candidate.is_dir() else dest
-
-
-def _write_zip_member(
-    zf: zipfile.ZipFile, info: zipfile.ZipInfo, dest: Path, *, total: int
-) -> tuple[str | None, int, dict[str, Any] | None]:
-    """Write one zip file entry. Returns (rel, new_total, fail_check)."""
-    if info.is_dir():
-        return None, total, None
-    rel = _safe_zip_relpath(info.filename)
-    if not rel:
-        return None, total, _zip_check("path", False, "unsafe_path", info.filename)
-    if not _zip_entry_allowed(rel):
-        return None, total, _zip_check("files", False, "disallowed_file", rel)
-    nxt = total + int(info.file_size or 0)
-    if nxt > _MAX_USER_SKILL_ZIP_UNCOMPRESSED:
-        return None, total, _zip_check("uncompressed", False, "uncompressed_too_large")
-    target = dest / rel
-    target.parent.mkdir(parents=True, exist_ok=True)
-    with zf.open(info, "r") as src, target.open("wb") as out:
-        shutil.copyfileobj(src, out)
-    return rel, nxt, None
-
-
-def _extract_user_skill_zip(raw: bytes, dest: Path) -> tuple[Path | None, list[dict[str, Any]]]:
-    """Safely extract a skill pack zip → (pack_dir, scan_checks)."""
-    checks: list[dict[str, Any]] = []
-    if len(raw) > _MAX_USER_SKILL_ZIP_BYTES:
-        checks.append(
-            _zip_check("size", False, "zip_too_large", f"max {_MAX_USER_SKILL_ZIP_BYTES} bytes")
-        )
-        return None, checks
-    checks.append(_zip_check("size", True, "zip_size_ok"))
-
-    try:
-        zf = zipfile.ZipFile(io.BytesIO(raw))
-    except zipfile.BadZipFile:
-        checks.append(_zip_check("format", False, "not_a_zip"))
-        return None, checks
-
-    written: list[str] = []
-    total = 0
-    with zf:
-        checks.append(_zip_check("format", True, "zip_format_ok"))
-        for info in zf.infolist():
-            rel, total, fail = _write_zip_member(zf, info, dest, total=total)
-            if fail:
-                checks.append(fail)
-                return None, checks
-            if rel:
-                written.append(rel)
-
-    if not written:
-        checks.append(_zip_check("files", False, "empty_zip"))
-        return None, checks
-    checks.append(_zip_check("path", True, "paths_ok"))
-
-    pack_dir = _resolve_extracted_pack_dir(dest, written)
-    has_meta = any((pack_dir / n).is_file() for n in _META_NAMES)
-    skill_md = _skill_md_path(pack_dir)
-    has_body = skill_md is not None
-    fm_ok = False
-    if has_body and not has_meta and skill_md is not None:
-        try:
-            fm, body = _split_skill_md_frontmatter(skill_md.read_text(encoding="utf-8"))
-            fm_ok = bool(body) and _meta_from_agent_skill_frontmatter(
-                fm, folder=pack_dir.name
-            ) is not None
-        except Exception:
-            fm_ok = False
-    meta_ok = has_meta or fm_ok
-    checks.append(
-        _zip_check(
-            "meta",
-            meta_ok,
-            "meta_present"
-            if has_meta
-            else ("meta_from_skill_frontmatter" if fm_ok else "meta_missing"),
-        )
-    )
-    checks.append(
-        _zip_check("skill_md", has_body, "skill_md_present" if has_body else "skill_md_missing")
-    )
-    if not meta_ok or not has_body:
-        return None, checks
-    return pack_dir, checks
-
-
-def _read_pack_meta(pack_dir: Path) -> dict[str, Any] | None:
-    for name in _META_NAMES:
-        p = pack_dir / name
-        if not p.is_file():
-            continue
-        meta = _read_json_file(p)
-        if meta:
-            return meta
-    skill_md = _skill_md_path(pack_dir)
-    if not skill_md:
-        return None
-    try:
-        fm, body = _split_skill_md_frontmatter(skill_md.read_text(encoding="utf-8"))
-    except Exception:
-        return None
-    if not body:
-        return None
-    return _meta_from_agent_skill_frontmatter(fm, folder=pack_dir.name)
-
-
-def _read_pack_skill_md(pack_dir: Path) -> tuple[str | None, str | None]:
-    """Return (body, error_label). Strips YAML frontmatter when present."""
-    skill_md = _skill_md_path(pack_dir)
-    if not skill_md:
-        return None, "skill_md_missing"
-    try:
-        raw = skill_md.read_text(encoding="utf-8")
-    except Exception:
-        return None, "skill_md_unreadable"
-    _fm, body = _split_skill_md_frontmatter(raw)
-    if not body:
-        return None, "prompt_positive_required"
-    return body, None
-
-
-def _pack_display_fields(meta: dict[str, Any], *, local: str) -> tuple[str, str, str]:
-    locales = meta.get("locales") if isinstance(meta.get("locales"), dict) else {}
-    loc = _locale_pick(locales) if locales else {}
-    display = str(
-        loc.get("displayName")
-        or loc.get("display_name")
-        or meta.get("displayName")
-        or meta.get("display_name")
-        or meta.get("title")
-        or meta.get("name")
-        or local
-    ).strip() or local
-    description = str(
-        loc.get("description")
-        or meta.get("description")
-        or meta.get("when_to_use")
-        or meta.get("whenToUse")
-        or ""
-    ).strip()
-    when = str(meta.get("when_to_use") or meta.get("whenToUse") or description).strip()
-    return display, description, when
-
-
-def _pack_logo_url(meta: dict[str, Any]) -> str | None:
-    logo_raw = str(meta.get("logo") or "").strip()
-    if logo_raw.startswith(("http://", "https://", "data:")):
-        return logo_raw
-    return None
-
-
-def _parse_user_skill_pack_dir(
-    pack_dir: Path,
-) -> tuple[dict[str, Any] | None, list[str]]:
-    """Parse pack dir → upsert payload (user namespace) + validation errors."""
-    meta = _read_pack_meta(pack_dir)
-    if not meta:
-        return None, ["meta_invalid"]
-    body, body_err = _read_pack_skill_md(pack_dir)
-    if body_err or not body:
-        return None, [body_err or "prompt_positive_required"]
-
-    folder = pack_dir.name
-    key_raw = str(
-        meta.get("skill_key") or meta.get("skillKey") or meta.get("name") or folder
-    ).strip()
-    ns_prefix, local = split_namespace_key(key_raw)
-    local = _slug_local_key((local or key_raw or folder).strip().lower() or folder)
-    errs: list[str] = []
-    if ns_prefix in (NS_CORE, NS_EXT):
-        errs.append(f"user_skill_cannot_use_{ns_prefix}_namespace")
-
-    skill_key = qualify_skill_key(NS_USER, local)
-    display, description, when = _pack_display_fields(meta, local=local)
-    pack_label, _ver = _parse_pack_version(
-        meta.get("version") or meta.get("pack_version") or meta.get("packVersion") or 1
-    )
-    errs.extend(
-        validate_skill_meta(
-            {
-                "skill_key": skill_key,
-                "name": display,
-                "prompt_positive": body,
-                "allowed_resources": ["tools"],
-                "namespace": NS_USER,
-                "input_schema": meta.get("input_schema") or meta.get("inputSchema"),
-                "output_schema": meta.get("output_schema") or meta.get("outputSchema"),
-            },
-            source=SOURCE_ADMIN,
-        )
-    )
-    if errs:
-        return None, errs
-
-    neg = str(meta.get("prompt_negative") or meta.get("promptNegative") or "").strip()
-    return {
-        "skillKey": skill_key,
-        "name": display,
-        "description": description or None,
-        "whenToUse": when or None,
-        "promptPositive": body,
-        "promptNegative": neg or None,
-        "category": str(meta.get("category") or "custom").strip() or "custom",
-        "logo": _pack_logo_url(meta),
-        "packVersion": pack_label,
-        "triggers": meta.get("triggers") or [{"type": "keyword", "value": display}],
-        "enabled": True,
-    }, []
-
-
-def _conflict_card(row: Any, *, mine: bool) -> dict[str, Any]:
-    pub = _pub(row)
-    return {
-        "id": int(pub.get("id") or 0),
-        "name": str(pub.get("name") or ""),
-        "skillKey": pub.get("skillKey"),
-        "packVersion": pub.get("packVersion") or str(pub.get("version") or "1"),
-        "updatedAt": float(_row_get(row, "updated_at") or 0) or None,
-        "useCount": 0,
-        "mine": mine,
-    }
-
-
-def _fetch_owned_skill_by_key(session: Session, *, uid: str, key: str) -> Any | None:
-    if not key:
-        return None
-    return crud.get_owned_design_skill_by_key(
-        session=session, owner_user_id=uid, skill_key=key
-    )
-
-
-def _fetch_owned_skill_by_name(session: Session, *, uid: str, name_l: str) -> Any | None:
-    if not name_l:
-        return None
-    return crud.get_owned_design_skill_by_name_lower(
-        session=session, owner_user_id=uid, name_lower=name_l
-    )
-
-
-def _fetch_foreign_skill_by_key(session: Session, *, uid: str, key: str) -> Any | None:
-    if not key:
-        return None
-    other = crud.get_design_skill_by_key(session=session, skill_key=key)
-    if not other:
-        return None
-    owner = str(_row_get(other, "owner_user_id") or "").strip()
-    if not owner or owner == uid:
-        return None
-    return other
-
-
-def _find_owned_skill_conflict(
-    *, user_id: str, skill_key: str, name: str
-) -> dict[str, Any] | None:
-    """Existing mine skill with same key/name, or foreign key collision."""
-    uid = str(user_id or "").strip()
-    key = str(skill_key or "").strip()
-    name_l = str(name or "").strip().lower()
-    ensure_design_skills()
-    with Session(core_db.engine) as session:
-        row = _fetch_owned_skill_by_key(session, uid=uid, key=key)
-        if not row:
-            row = _fetch_owned_skill_by_name(session, uid=uid, name_l=name_l)
-        if row:
-            return _conflict_card(row, mine=True)
-        foreign = _fetch_foreign_skill_by_key(session, uid=uid, key=key)
-        if foreign:
-            return _conflict_card(foreign, mine=False)
-    return None
-
-
-def _apply_pack_version(*, uid: str, skill_id: int, pack_ver: str | None) -> None:
-    if not pack_ver or skill_id <= 0:
-        return
-    try:
-        with Session(core_db.engine) as session:
-            crud.update_design_skill_pack_version(
-                session=session,
-                item_id=skill_id,
-                owner_user_id=uid,
-                pack_version=pack_ver,
-            )
-    except Exception:
-        logger.exception("pack_version update failed")
-
-
-def import_end_user_skill_zip(
-    *,
-    user_id: str,
-    filename: str,
-    raw: bytes,
-    overwrite: bool = False,
-) -> dict[str, Any]:
-    """Import a skill pack zip for the signed-in user (scan → upsert)."""
-    uid = str(user_id or "").strip()
-    fname = str(filename or "skill.zip").strip() or "skill.zip"
-    checks: list[dict[str, Any]] = []
-
-    if not fname.lower().endswith(".zip"):
-        checks.append(_zip_check("ext", False, "need_zip"))
-        return _import_result(
-            status="rejected", file_name=fname, checks=checks, errors=["need_zip"], scan_ok=False
-        )
-    checks.append(_zip_check("ext", True, "zip_extension"))
-
-    if not raw:
-        checks.append(_zip_check("empty", False, "empty_file"))
-        return _import_result(
-            status="rejected", file_name=fname, checks=checks, errors=["empty_file"], scan_ok=False
-        )
-
-    tmp = Path(tempfile.mkdtemp(prefix="skill-zip-"))
-    try:
-        pack_dir, extract_checks = _extract_user_skill_zip(raw, tmp)
-        checks.extend(extract_checks)
-        if not pack_dir:
-            errs = [c["label"] for c in extract_checks if not c.get("ok")]
-            return _import_result(
-                status="rejected", file_name=fname, checks=checks, errors=errs, scan_ok=False
-            )
-
-        payload, parse_errs = _parse_user_skill_pack_dir(pack_dir)
-        if parse_errs or not payload:
-            checks.append(_zip_check("meta_valid", False, "meta_invalid"))
-            return _import_result(
-                status="rejected",
-                file_name=fname,
-                checks=checks,
-                errors=parse_errs or ["meta_invalid"],
-                scan_ok=False,
-            )
-        checks.append(_zip_check("meta_valid", True, "meta_ok"))
-
-        conflict = _find_owned_skill_conflict(
-            user_id=uid,
-            skill_key=str(payload.get("skillKey") or ""),
-            name=str(payload.get("name") or ""),
-        )
-        if conflict and not conflict.get("mine"):
-            checks.append(_zip_check("ownership", False, "skill_key_taken_other"))
-            return _import_result(
-                status="rejected",
-                file_name=fname,
-                checks=checks,
-                errors=["skill_key_taken_other"],
-                existing=conflict,
-                scan_ok=False,
-            )
-        if conflict and conflict.get("mine") and not overwrite:
-            return _import_result(
-                status="exists",
-                file_name=fname,
-                checks=checks,
-                existing=conflict,
-                scan_ok=True,
-            )
-        if conflict and conflict.get("mine"):
-            payload["id"] = int(conflict["id"])
-
-        pack_ver = str(payload.pop("packVersion", "") or "").strip() or None
-        item = upsert_end_user_skill(user_id=uid, payload=payload)
-        _apply_pack_version(uid=uid, skill_id=int(item.get("id") or 0), pack_ver=pack_ver)
-        if pack_ver:
-            item["packVersion"] = pack_ver
-        return _import_result(
-            status="ok", file_name=fname, checks=checks, item=item, scan_ok=True
-        )
-    finally:
-        shutil.rmtree(tmp, ignore_errors=True)
