+++ apps/api/app/services/design/prompts/skill_store/__init__.py
@@ -1,0 +1,286 @@
+"""Runtime skills for lc_design — pluggable catalog + need_skills + triggers + files.
+
+Namespaces (conflict isolation):
+  - core  — system seed skills (source=seed); bare keys kept for BC
+  - ext   — file packs under ``.agents/skills`` + data/design_skills (source=file)
+  - user  — admin / user-extension skills (source=admin); keys use ``user.<local>``
+
+Also:
+  - ACL: preferred_tools + allowed_resources; user skills never reopen unrestricted surface
+  - I/O: pack/admin meta validation; need_skills pin/args vs input_schema; ops vs output_schema
+  - Version: integer bump + pack_version + design_skill_revision snapshots; pin via ``key@N``
+  - Hot reload: optional mtime watcher on seed JSON + file packs
+  - Formats: product ``_meta.json``+``SKILL.md``, or ``SKILL.md`` with YAML frontmatter
+"""
+from __future__ import annotations
+
+from .constants import (
+    MAX_SKILL_DETAIL_CHARS,
+    NS_CORE,
+    NS_EXT,
+    NS_USER,
+    RETIRED_NEED_PROMPT_KINDS,
+    SOURCE_ADMIN,
+    SOURCE_FILE,
+    SOURCE_SEED,
+    _ALWAYS_ALLOW_OPS,
+    _INTERNAL_RESOURCE_KINDS,
+    _MAX_USER_SKILL_ZIP_BYTES,
+    _MAX_USER_SKILL_ZIP_UNCOMPRESSED,
+    _META_NAMES,
+    _NS_KEY_RE,
+    _PIN_RE,
+    _PROTECTED_FROM_FILE,
+    _PROTECTED_FROM_SEED,
+    _SKILL_MD_NAMES,
+    _SLUG_RE,
+    _SOURCE_TO_NS,
+    _VALID_NAMESPACES,
+    _ZIP_LOGO_EXTS,
+)
+from .ensure import (
+    _SEED_SKILL_BODY_MARKERS,
+    _allowed_resources_json,
+    _bump_unchanged_seed_skill_bodies,
+    _norm_skill_body,
+    _preferred_json,
+    _schema_json,
+    _skills_disk_signature,
+    _triggers_json,
+    _upsert_owned_skill,
+    ensure_design_skills,
+    reload_skills_if_disk_changed,
+    start_skills_hot_reload,
+    stop_skills_hot_reload,
+)
+from .keys import (
+    _normalize_namespace,
+    _normalize_source,
+    _slug_local_key,
+    parse_skill_pin,
+    qualify_skill_key,
+    skill_kind_for_namespace,
+    split_namespace_key,
+)
+from .pack_io import (
+    _CORE_RESERVED_KEYS,
+    _SEED,
+    _SEED_BY_KEY,
+    _agents_skills_dir,
+    _file_skills_dir,
+    _file_skills_dirs,
+    _load_file_skills,
+    _load_pack_dir,
+    _load_skills_seed,
+    _locale_pick,
+    _meta_from_agent_skill_frontmatter,
+    _pack_has_product_meta,
+    _parse_pack_version,
+    _read_json_file,
+    _repo_root,
+    _resolve_pack_logo,
+    _skill_item_from_parts,
+    _skill_md_path,
+    _skills_seed_path,
+    _split_skill_md_frontmatter,
+    _unquote_yaml_scalar,
+)
+from .runtime import (
+    _apply_mutex,
+    _csv_has,
+    _iter_skills_for_keys,
+    _load_skill_revision_snapshot,
+    _load_user_skill_prefs,
+    _normalize_loaded_skill_keys,
+    _parse_allowed_resources,
+    _parse_preferred_tools,
+    _parse_triggers,
+    _pub,
+    _row_get,
+    _rule_matches,
+    filter_need_resources_by_skill_acl,
+    filter_ops_by_skill_allowlist,
+    filter_ops_by_skill_output_schema,
+    format_skills_catalog,
+    format_skills_details,
+    format_skills_details_checked,
+    invalidate_skill_key_cache,
+    list_runtime_skills,
+    normalize_need_skills,
+    parse_need_skills_with_pins,
+    preferred_tools_allowlist,
+    reset_skills_ready_for_tests,
+    resolve_storage_skill_key,
+    resolve_triggered_skill_keys,
+    runtime_skill_keys,
+    save_skill_revision,
+    skill_resource_allowlist,
+)
+from .schema import (
+    _parse_json_object,
+    validate_against_schema,
+    validate_skill_io_schema,
+    validate_skill_meta,
+)
+from .user_skills import (
+    _apply_pack_version,
+    _conflict_card,
+    _extract_user_skill_zip,
+    _fetch_foreign_skill_by_key,
+    _fetch_owned_skill_by_key,
+    _fetch_owned_skill_by_name,
+    _find_owned_skill_conflict,
+    _import_result,
+    _pack_display_fields,
+    _pack_logo_url,
+    _parse_user_skill_pack_dir,
+    _picker_card,
+    _read_pack_meta,
+    _read_pack_skill_md,
+    _resolve_extracted_pack_dir,
+    _safe_zip_relpath,
+    _write_zip_member,
+    _zip_check,
+    _zip_entry_allowed,
+    delete_end_user_skill,
+    import_end_user_skill_zip,
+    list_my_skills,
+    list_skills_for_manage,
+    list_skills_for_picker,
+    resolve_accessible_skill_keys,
+    set_user_skill_enabled,
+    upsert_end_user_skill,
+)
+
+__all__ = [
+    # constants
+    "SOURCE_SEED",
+    "SOURCE_ADMIN",
+    "SOURCE_FILE",
+    "NS_CORE",
+    "NS_EXT",
+    "NS_USER",
+    "RETIRED_NEED_PROMPT_KINDS",
+    "MAX_SKILL_DETAIL_CHARS",
+    "_ALWAYS_ALLOW_OPS",
+    "_PROTECTED_FROM_SEED",
+    "_PROTECTED_FROM_FILE",
+    "_VALID_NAMESPACES",
+    "_SOURCE_TO_NS",
+    "_META_NAMES",
+    "_SKILL_MD_NAMES",
+    "_NS_KEY_RE",
+    "_PIN_RE",
+    "_INTERNAL_RESOURCE_KINDS",
+    "_SLUG_RE",
+    "_MAX_USER_SKILL_ZIP_BYTES",
+    "_MAX_USER_SKILL_ZIP_UNCOMPRESSED",
+    "_ZIP_LOGO_EXTS",
+    # keys
+    "split_namespace_key",
+    "qualify_skill_key",
+    "skill_kind_for_namespace",
+    "parse_skill_pin",
+    "_normalize_namespace",
+    "_normalize_source",
+    "_slug_local_key",
+    # schema
+    "validate_skill_meta",
+    "validate_against_schema",
+    "validate_skill_io_schema",
+    "_parse_json_object",
+    # pack_io
+    "_SEED",
+    "_SEED_BY_KEY",
+    "_CORE_RESERVED_KEYS",
+    "_load_file_skills",
+    "_load_pack_dir",
+    "_load_skills_seed",
+    "_parse_pack_version",
+    "_meta_from_agent_skill_frontmatter",
+    "_split_skill_md_frontmatter",
+    "_skill_md_path",
+    "_unquote_yaml_scalar",
+    "_skills_seed_path",
+    "_repo_root",
+    "_agents_skills_dir",
+    "_file_skills_dir",
+    "_file_skills_dirs",
+    "_pack_has_product_meta",
+    "_read_json_file",
+    "_locale_pick",
+    "_resolve_pack_logo",
+    "_skill_item_from_parts",
+    # runtime
+    "list_runtime_skills",
+    "format_skills_catalog",
+    "format_skills_details",
+    "format_skills_details_checked",
+    "filter_ops_by_skill_allowlist",
+    "filter_ops_by_skill_output_schema",
+    "filter_need_resources_by_skill_acl",
+    "parse_need_skills_with_pins",
+    "normalize_need_skills",
+    "resolve_triggered_skill_keys",
+    "resolve_storage_skill_key",
+    "preferred_tools_allowlist",
+    "skill_resource_allowlist",
+    "runtime_skill_keys",
+    "invalidate_skill_key_cache",
+    "reset_skills_ready_for_tests",
+    "save_skill_revision",
+    "_pub",
+    "_row_get",
+    "_apply_mutex",
+    "_rule_matches",
+    "_parse_preferred_tools",
+    "_parse_allowed_resources",
+    "_parse_triggers",
+    "_csv_has",
+    "_load_user_skill_prefs",
+    "_load_skill_revision_snapshot",
+    "_normalize_loaded_skill_keys",
+    "_iter_skills_for_keys",
+    # ensure
+    "ensure_design_skills",
+    "start_skills_hot_reload",
+    "stop_skills_hot_reload",
+    "reload_skills_if_disk_changed",
+    "_upsert_owned_skill",
+    "_skills_disk_signature",
+    "_SEED_SKILL_BODY_MARKERS",
+    "_norm_skill_body",
+    "_bump_unchanged_seed_skill_bodies",
+    "_triggers_json",
+    "_preferred_json",
+    "_allowed_resources_json",
+    "_schema_json",
+    # user_skills
+    "list_skills_for_picker",
+    "list_skills_for_manage",
+    "list_my_skills",
+    "set_user_skill_enabled",
+    "resolve_accessible_skill_keys",
+    "upsert_end_user_skill",
+    "delete_end_user_skill",
+    "import_end_user_skill_zip",
+    "_picker_card",
+    "_zip_check",
+    "_import_result",
+    "_zip_entry_allowed",
+    "_safe_zip_relpath",
+    "_resolve_extracted_pack_dir",
+    "_write_zip_member",
+    "_extract_user_skill_zip",
+    "_read_pack_meta",
+    "_read_pack_skill_md",
+    "_pack_display_fields",
+    "_pack_logo_url",
+    "_parse_user_skill_pack_dir",
+    "_conflict_card",
+    "_fetch_owned_skill_by_key",
+    "_fetch_owned_skill_by_name",
+    "_fetch_foreign_skill_by_key",
+    "_find_owned_skill_conflict",
+    "_apply_pack_version",
+]
