+++ apps/api/data/design_knowledge_seed.json
@@ -67,7 +67,7 @@
       "skill_categories": "validate,layout,refine",
       "sort_order": 20,
       "when_to_use": "网站、App、后台界面组件与栅格",
-      "body": "网格：8px/4px 基底；页面/卡片边距尽量统一。\n组件：按钮/输入/开关/标签/弹窗全局一套；按钮分主/次/文字/危险；卡片圆角/描边统一。\n图标：默认优先 create_svg 自绘；复杂/写实可用 create_image（用户不要生图则只用矢量）；线粗统一；尺寸档 16/20/24；线性或面性勿混。\n色板与字号层级硬检见 VISUAL_GATES / FE QA。\n图层：背景→容器→组件→文字；配图用 create_image（PNG），勿留空白/纯色矩形占位。"
+      "body": "网格：8px/4px 基底；页面/卡片边距尽量统一。\n组件：按钮/输入/开关/标签/弹窗全局一套；按钮分主/次/文字/危险；卡片圆角/描边统一。\n图标：优先几何 shape + boolean_op / 钢笔构建可编辑矢量；简单单路径可用 create_svg；复杂写实才 create_image（用户不要生图则只用矢量）；线粗统一；尺寸档 16/20/24；线性或面性勿混。\n动效：加载/成功/空状态循环用 create_lottie + genPrompt；静态图标不要用 Lottie。\n色板与字号层级硬检见 VISUAL_GATES / FE QA。\n图层：背景→容器→组件→文字；配图用 create_image（PNG），勿留空白/纯色矩形占位。"
     },
     {
       "kind": "icon",
@@ -75,8 +75,8 @@
       "scenes": "website,mobile,image",
       "skill_categories": "layout,validate,refine",
       "sort_order": 25,
-      "when_to_use": "画 UI 图标、功能入口、导航符号，或用户要求图标风格统一时",
-      "body": "【风格统一】同一界面只选一种：线性 / 面性 / 双色；勿混 3D、手绘、渐变浮雕、写实插画图标。\n【绘制手段·默认】功能图标优先 create_svg：由你自己写简洁 args.svg（viewBox 0 0 24 24），落成原生 SVG 节点；不要抄预设图标库/美学配方；SVG 表达不了的复杂/写实图标才用 create_image + genPrompt（PNG）；禁止空圆/色块冒充图标；少用 create_shape type=path 临场乱编超长 d。\n【尺寸档】常用 16 / 20 / 24 / 32；同一层级尺寸一致；触控旁图标可 20～24，列表辅图标宜 16～20。\n【线粗与端点】线性图标线粗全局统一（如 1.5 或 2px）；圆角端点或直角端点二选一；关键形闭合完整，忌断续碎线。\n【网格与对齐】放在正方形安全区，光学居中（视觉中心略上移可接受）；外轮廓对齐像素格；复杂图标先几何化再细节。\n【识别性】轮廓剪影一眼可读；隐喻通用（搜索=放大镜、设置=齿轮）；细节服从整体，小尺寸少内部分割。\n【颜色】默认单色（跟随文字色/主色）；状态可用成功/警告/错误色；彩色图标强调色通常 ≤2。\n【禁区】勿用碎矩形/斜线堆假图标；勿用不准确超长 path 冒充齿轮（改用 create_svg 自绘简洁符号）；勿在后台堆装饰性插画当功能图标。\n【用户指定】用户明确不要生图/只要矢量时：只用 create_svg 或几何 shape，不要 create_image。"
+      "when_to_use": "画 UI 图标、功能入口、导航符号、小程序图标套装，或用户要求图标风格统一时",
+      "body": "【风格统一】同一套只选一种：线性 / 面性 / 双色；勿混 3D、手绘、渐变浮雕、写实插画图标。\n【绘制手段·构造优先】功能图标优先可编辑矢量构造，不要只会 dump create_svg：\n1) 几何：create_shape（rect/ellipse/circle/triangle/polygon/star）搭外形；\n2) 布尔：boolean_op（union 合并 / subtract 挖空 / intersect / exclude）做扣洞、拼合（如放大镜=圆∪柄、月亮=大圆 subtract 小圆）；\n3) 钢笔：create_shape shapeType=pen + path（需要闭合填色时 closed）；有机曲线、定制剪影用钢笔，勿用碎矩形冒充；\n4) create_svg / create_icon：仅适合极简单路径符号，或布尔/钢笔会更啰嗦时；\n5) create_image：仅复杂写实/插画风图标，且用户未禁止生图。\n禁止空圆/色块冒充；禁止临场乱编超长 path d 冒充齿轮。\n【套装】如「旅游小程序一整套图标」：先定统一线粗/圆角/填色语言，再按网格落多个同尺寸单元格；每格一枚图标+短标签；隐喻通用（首页、行程、景点、酒店、机票、攻略、收藏、我的）。\n【尺寸档】常用 16 / 20 / 24 / 32；同一层级尺寸一致；触控旁 20～24，列表辅图标 16～20。\n【线粗与端点】线性图标线粗全局统一（如 1.5 或 2px）；圆角端点或直角端点二选一；关键形闭合完整。\n【网格与对齐】正方形安全区，光学居中；外轮廓对齐像素格；复杂图标先几何化再细节。\n【识别性】轮廓剪影一眼可读；细节服从整体，小尺寸少内部分割。\n【颜色】默认单色；状态可用成功/警告/错误色；彩色强调色通常 ≤2。\n【禁区】勿用碎矩形/斜线堆假图标；勿在后台堆装饰性插画当功能图标。\n【用户指定】不要生图/只要矢量：只用 shape + boolean_op + pen / create_svg，不要 create_image。"
     },
     {
       "kind": "techniques",
@@ -94,7 +94,7 @@
       "skill_categories": "layout",
       "sort_order": 30,
       "when_to_use": "字体族选择、对齐方式不清晰时；艺术字 vs 真文字选型",
-      "body": "字号跨度硬检见 VISUAL_GATES / FE hierarchy。\n字体族：同一设计通常 ≤2；UI 禁用书法/路径美术字作正文。\n对齐：正文左对齐为主；居中仅短标题/英雄区。\n【艺术字 vs 真文字】特效字形（笔触/分色/立体等）→ create_image 作图；需要可编辑排版的字 → create_text。由模型按效果判断，勿按文案品类写死。"
+      "body": "字号跨度硬检见 VISUAL_GATES / FE hierarchy。\n字体族：同一设计通常 ≤2；UI 禁用书法/路径美术字作正文。\n对齐：正文左对齐为主；居中仅短标题/英雄区。\n【艺术字 vs 真文字 · ~90% 相似门槛】对比「需要的标题气质」（USER_PROMPT + 参考图）与 Available fonts：\n- 仅当某字库面相约 **≥90%** 相似（字重/结构/气质/语种）才用 create_text + fontFamily；\n- 低于门槛——尤其主标题 / 英雄大字 / 海报 display——用 create_image + genPrompt + letteringText；\n- 勿把「书法感 / 手写 / 国潮 / 艺术字」写死映射到某一书法字体；勿发明字库外字体名。\n正文/UI 可读性优先字库；特效字形（分色/立体/定制笔触）按效果判断，勿按文案品类写死。"
     },
     {
       "kind": "ia",
@@ -202,7 +202,7 @@
       "skill_categories": "layout,refine,validate",
       "sort_order": 60,
       "when_to_use": "插画、角色、贴纸、手绘风图形，或用户要求板绘/厚涂/线稿上色时",
-      "body": "【媒介】板绘=数字绘画成稿，不是 UI 线框；少表单/导航壳；需要界面时分开：画面主体用绘制，控件用 UI 规范。\n【流程铁律】任何画风都先大体结构：构图剪影 → 比例结构 → 大色块 → 光影 → 细节风格；禁止先画五官/花纹/高光。详见「绘画流程」知识。\n【形与透视】主体剪影清晰可读；人物/器物结构准；透视统一（一点/两点），地平线稳定。\n【光影】明确主光源方向；暗部有环境反光；厚涂可柔边过渡，赛璐璐宜分面干净；勿到处均匀高光。\n【色彩】定主色调再铺；阴影偏冷或同色相降明度；避免脏灰（补色相混过量）；重点色小面积点缀。\n【笔触】同一张画笔触语言统一（平滑厚涂 / 可见笔触 / 扁平矢量）；线稿线粗有主次；勾线闭合完整，少碎线堆材质。\n【图层意识】背景→主体→前景；边缘干净，主体与背景分离度够。\n【输出】边缘锐利可裁；透明底贴纸留安全边；禁止用杂乱小矩形冒充笔触或未完成草稿当交付。"
+      "body": "【媒介】板绘=数字绘画成稿，不是 UI 线框；少表单/导航壳；需要界面时分开：画面主体用绘制，控件用 UI 规范。\n【画布笔触】板绘/线稿用 create_shape shapeType=pencil（tip 笔尖 stamp，不是圆拼贴）：path 用 M/L 折线；pathPressure csv（0.05–1，点数=折线点数）；brushStyle tip id（solid|pencil-hb|soft|fountain|calligraphy|brushpen|marker|highlighter|chalk|charcoal|bristle|airbrush|watercolor|needle|bold）；brushHardness 0–100 控笔尖软硬（默认约 80）；pressureEnabled 控是否吃压感（有 pathPressure 时默认开）。多笔表达，统一一种 tip。\n【流程铁律】任何画风都先大体结构：构图剪影 → 比例结构 → 大色块 → 光影 → 细节风格；禁止先画五官/花纹/高光。详见「绘画流程」知识。\n【形与透视】主体剪影清晰可读；人物/器物结构准；透视统一（一点/两点），地平线稳定。\n【光影】明确主光源方向；暗部有环境反光；厚涂可柔边过渡，赛璐璐宜分面干净；勿到处均匀高光。\n【色彩】定主色调再铺；阴影偏冷或同色相降明度；避免脏灰（补色相混过量）；重点色小面积点缀。\n【笔触】同一张画笔触语言统一（平滑厚涂 / 可见笔触 / 扁平矢量）；线稿线粗有主次；勾线闭合完整，少碎线堆材质。\n【图层意识】背景→主体→前景；边缘干净，主体与背景分离度够。\n【输出】边缘锐利可裁；透明底贴纸留安全边；禁止用杂乱小矩形冒充笔触或未完成草稿当交付。"
     },
     {
       "kind": "paint_process",
